saving new profile for admin view along with deletion of old admin assets
