+++ shift-scan/public/sw.js
@@ -1,2 +1,2 @@
-if(!self.define){let e,a={};const i=(i,s)=>(i=new URL(i+".js",s).href,a[i]||new Promise((a=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=a,document.head.appendChild(e)}else e=i,importScripts(i),a()})).then((()=>{let e=a[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e})));self.define=(s,c)=>{const r=e||("document"in self?document.currentScript.src:"")||location.href;if(a[r])return;let n={};const t=e=>i(e,r),d={module:{uri:r},exports:n,require:t};a[r]=Promise.all(s.map((e=>d[e]||t(e)))).then((e=>(c(...e),n)))}}define(["./workbox-31da5539"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/Settings.svg",revision:"ebe8b6a8bff9bd2fe8669ad036e8edf8"},{url:"/_next/static/chunks/1183-1edb584d8d091e42.js",revision:"1edb584d8d091e42"},{url:"/_next/static/chunks/1401-0de5bc94198b3c9a.js",revision:"0de5bc94198b3c9a"},{url:"/_next/static/chunks/1692-cc16fa14e0e5bbf3.js",revision:"cc16fa14e0e5bbf3"},{url:"/_next/static/chunks/1791-a4cd8ef707d3c1ff.js",revision:"a4cd8ef707d3c1ff"},{url:"/_next/static/chunks/2019-244f4105c4691c0c.js",revision:"244f4105c4691c0c"},{url:"/_next/static/chunks/2057-668deaa0eb77c6a4.js",revision:"668deaa0eb77c6a4"},{url:"/_next/static/chunks/2170a4aa-8c96009f60e20411.js",revision:"8c96009f60e20411"},{url:"/_next/static/chunks/2232-3f04a14cda4cd5a1.js",revision:"3f04a14cda4cd5a1"},{url:"/_next/static/chunks/2403.ca107014ddba66de.js",revision:"ca107014ddba66de"},{url:"/_next/static/chunks/2441-9ad2299f689922c4.js",revision:"9ad2299f689922c4"},{url:"/_next/static/chunks/2927-0615eaa2f61c8eef.js",revision:"0615eaa2f61c8eef"},{url:"/_next/static/chunks/3008-31aa8928ac008967.js",revision:"31aa8928ac008967"},{url:"/_next/static/chunks/3031-e0cbd25419c08505.js",revision:"e0cbd25419c08505"},{url:"/_next/static/chunks/3068-9537e5f9d379785b.js",revision:"9537e5f9d379785b"},{url:"/_next/static/chunks/3328-1e73249e263f1cb2.js",revision:"1e73249e263f1cb2"},{url:"/_next/static/chunks/3343-cddd99d48f629228.js",revision:"cddd99d48f629228"},{url:"/_next/static/chunks/4095-8d931f3f0218645a.js",revision:"8d931f3f0218645a"},{url:"/_next/static/chunks/4162-f6605c3f153eb555.js",revision:"f6605c3f153eb555"},{url:"/_next/static/chunks/4387-3676ac23a2682f13.js",revision:"3676ac23a2682f13"},{url:"/_next/static/chunks/460-8f004685eb760a9f.js",revision:"8f004685eb760a9f"},{url:"/_next/static/chunks/4610-9b372e525bb28e3c.js",revision:"9b372e525bb28e3c"},{url:"/_next/static/chunks/4692-79463bd8f0d76a0e.js",revision:"79463bd8f0d76a0e"},{url:"/_next/static/chunks/4bd1b696-88e37c56edce0f59.js",revision:"88e37c56edce0f59"},{url:"/_next/static/chunks/5152-9e6b4080454ffa3c.js",revision:"9e6b4080454ffa3c"},{url:"/_next/static/chunks/5163-a9ef2a6bf3806903.js",revision:"a9ef2a6bf3806903"},{url:"/_next/static/chunks/5166-22f32cf0bc0e8fc2.js",revision:"22f32cf0bc0e8fc2"},{url:"/_next/static/chunks/52774a7f-9f526229934f791c.js",revision:"9f526229934f791c"},{url:"/_next/static/chunks/5493-002a49a6d796cca8.js",revision:"002a49a6d796cca8"},{url:"/_next/static/chunks/5751-70590a584bd12447.js",revision:"70590a584bd12447"},{url:"/_next/static/chunks/602-c9c2a774f75643ad.js",revision:"c9c2a774f75643ad"},{url:"/_next/static/chunks/6323-d1e244e417681874.js",revision:"d1e244e417681874"},{url:"/_next/static/chunks/6544-bcf8e533ec6b3434.js",revision:"bcf8e533ec6b3434"},{url:"/_next/static/chunks/6671-c6ce7faab2868fe3.js",revision:"c6ce7faab2868fe3"},{url:"/_next/static/chunks/6842-d6266f7a919f2655.js",revision:"d6266f7a919f2655"},{url:"/_next/static/chunks/6862-81cfc3fff123ab8a.js",revision:"81cfc3fff123ab8a"},{url:"/_next/static/chunks/6874-bf3fa41ffddf2fd7.js",revision:"bf3fa41ffddf2fd7"},{url:"/_next/static/chunks/7053-2689323dc7a6c8d9.js",revision:"2689323dc7a6c8d9"},{url:"/_next/static/chunks/707-3d319102f6468d5f.js",revision:"3d319102f6468d5f"},{url:"/_next/static/chunks/7082-c91642ee82e5320d.js",revision:"c91642ee82e5320d"},{url:"/_next/static/chunks/7111-32719b32e4f3138d.js",revision:"32719b32e4f3138d"},{url:"/_next/static/chunks/7253-b6a7c4b8ba2a4ccc.js",revision:"b6a7c4b8ba2a4ccc"},{url:"/_next/static/chunks/7765-7d2d9a5f429672fa.js",revision:"7d2d9a5f429672fa"},{url:"/_next/static/chunks/789-86a3b92415cd819c.js",revision:"86a3b92415cd819c"},{url:"/_next/static/chunks/7914-e7d7b2a8c8cab0d5.js",revision:"e7d7b2a8c8cab0d5"},{url:"/_next/static/chunks/8007-9d7908d793cc20bb.js",revision:"9d7908d793cc20bb"},{url:"/_next/static/chunks/822-0edc64ee089dcd0d.js",revision:"0edc64ee089dcd0d"},{url:"/_next/static/chunks/8436.fe8b1513f8e56bfa.js",revision:"fe8b1513f8e56bfa"},{url:"/_next/static/chunks/9313-eae8d054304ecf16.js",revision:"eae8d054304ecf16"},{url:"/_next/static/chunks/9337-ed07232187c0fdae.js",revision:"ed07232187c0fdae"},{url:"/_next/static/chunks/9373-12ddcc70f6722621.js",revision:"12ddcc70f6722621"},{url:"/_next/static/chunks/949-b28f693ac97fe61e.js",revision:"b28f693ac97fe61e"},{url:"/_next/static/chunks/9648-fecb89b91c985f68.js",revision:"fecb89b91c985f68"},{url:"/_next/static/chunks/9763-3b7d58f5c287c919.js",revision:"3b7d58f5c287c919"},{url:"/_next/static/chunks/app/(routes)/admins/clients/page-3fa8a7055592d9d1.js",revision:"3fa8a7055592d9d1"},{url:"/_next/static/chunks/app/(routes)/admins/cost-codes/page-5b613be8b88a8ed4.js",revision:"5b613be8b88a8ed4"},{url:"/_next/static/chunks/app/(routes)/admins/equipment/page-c4c7aa3006934c70.js",revision:"c4c7aa3006934c70"},{url:"/_next/static/chunks/app/(routes)/admins/forms/%5Bid%5D/page-b1975f3c1114c1fd.js",revision:"b1975f3c1114c1fd"},{url:"/_next/static/chunks/app/(routes)/admins/forms/create/page-d7d1c50fe54db441.js",revision:"d7d1c50fe54db441"},{url:"/_next/static/chunks/app/(routes)/admins/forms/edit/%5Bid%5D/page-04cfc5424283b722.js",revision:"04cfc5424283b722"},{url:"/_next/static/chunks/app/(routes)/admins/forms/page-c4569757416362f7.js",revision:"c4569757416362f7"},{url:"/_next/static/chunks/app/(routes)/admins/jobsites/page-c9d2fff499f01793.js",revision:"c9d2fff499f01793"},{url:"/_next/static/chunks/app/(routes)/admins/layout-cda8d903ae2ae668.js",revision:"cda8d903ae2ae668"},{url:"/_next/static/chunks/app/(routes)/admins/page-111bcb9c8973e94e.js",revision:"111bcb9c8973e94e"},{url:"/_next/static/chunks/app/(routes)/admins/personnel/page-826711ddb5f9a33d.js",revision:"826711ddb5f9a33d"},{url:"/_next/static/chunks/app/(routes)/admins/reports/page-7548cd12dc68da90.js",revision:"7548cd12dc68da90"},{url:"/_next/static/chunks/app/(routes)/admins/timesheets/page-0188d58e99d2beea.js",revision:"0188d58e99d2beea"},{url:"/_next/static/chunks/app/(routes)/break/page-e81008d07e2321ea.js",revision:"e81008d07e2321ea"},{url:"/_next/static/chunks/app/(routes)/clock/page-f693f2e9ce928a45.js",revision:"f693f2e9ce928a45"},{url:"/_next/static/chunks/app/(routes)/dashboard/clock-out/(components)/log/page-8e60956c7cc32cd2.js",revision:"8e60956c7cc32cd2"},{url:"/_next/static/chunks/app/(routes)/dashboard/clock-out/page-be4435e6e9c4e7b7.js",revision:"be4435e6e9c4e7b7"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/%5Bid%5D/page-8e5b5fc236df03ee.js",revision:"8e5b5fc236df03ee"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/log-new/page-3f6c261ac7be890d.js",revision:"3f6c261ac7be890d"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/page-78a72de45d1c6146.js",revision:"78a72de45d1c6146"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/edit-repair-details/%5Bid%5D/page-56d27e11be44efa6.js",revision:"56d27e11be44efa6"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/new-repair/page-492d857bb72f9d7d.js",revision:"492d857bb72f9d7d"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/page-bb69016e41d01914.js",revision:"bb69016e41d01914"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/projects/%5Bid%5D/page-a46164a20b3e6c25.js",revision:"a46164a20b3e6c25"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/%5Bid%5D/employee/%5BemployeeId%5D/page-120a6a015e972199.js",revision:"120a6a015e972199"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/%5Bid%5D/page-9d5a0f7593de0429.js",revision:"9d5a0f7593de0429"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/page-d56082030b4496ed.js",revision:"d56082030b4496ed"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/timecards/page-761165f83fc91545.js",revision:"761165f83fc91545"},{url:"/_next/static/chunks/app/(routes)/dashboard/page-4016bf3e76d83acb.js",revision:"4016bf3e76d83acb"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/add-equipment/page-1e648bfb9a300d32.js",revision:"1e648bfb9a300d32"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/add-jobsite/page-a89384fb40d47e71.js",revision:"a89384fb40d47e71"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/page-a0594747f8a8bc58.js",revision:"a0594747f8a8bc58"},{url:"/_next/static/chunks/app/(routes)/dashboard/switch-jobs/page-b4e84d58fdc7476c.js",revision:"b4e84d58fdc7476c"},{url:"/_next/static/chunks/app/(routes)/dashboard/tasco/page-3bb97b3aea503c41.js",revision:"3bb97b3aea503c41"},{url:"/_next/static/chunks/app/(routes)/dashboard/truckingAssistant/page-fc0534ab53dede76.js",revision:"fc0534ab53dede76"},{url:"/_next/static/chunks/app/(routes)/hamburger/changePassword/page-9f6a991581c388a4.js",revision:"9f6a991581c388a4"},{url:"/_next/static/chunks/app/(routes)/hamburger/inbox/formSubmission/%5Bid%5D/page-60c440af1d4621cb.js",revision:"60c440af1d4621cb"},{url:"/_next/static/chunks/app/(routes)/hamburger/inbox/page-9dc20b54139982dd.js",revision:"9dc20b54139982dd"},{url:"/_next/static/chunks/app/(routes)/hamburger/profile/page-882d71f23505fdd6.js",revision:"882d71f23505fdd6"},{url:"/_next/static/chunks/app/(routes)/information/page-89d7583d61422b6d.js",revision:"89d7583d61422b6d"},{url:"/_next/static/chunks/app/(routes)/signin/forgot-password/page-b33d9bbb15f5cbfe.js",revision:"b33d9bbb15f5cbfe"},{url:"/_next/static/chunks/app/(routes)/signin/new-password/page-db8596a24fb7b927.js",revision:"db8596a24fb7b927"},{url:"/_next/static/chunks/app/(routes)/signin/page-776eda5d900bfbb8.js",revision:"776eda5d900bfbb8"},{url:"/_next/static/chunks/app/(routes)/signin/signup/page-b9664185c91a1fdf.js",revision:"b9664185c91a1fdf"},{url:"/_next/static/chunks/app/(routes)/timesheets/page-525cb415f3827045.js",revision:"525cb415f3827045"},{url:"/_next/static/chunks/app/_not-found/page-9c84c29c80583905.js",revision:"9c84c29c80583905"},{url:"/_next/static/chunks/app/api/(admins)/contactInfo/%5Bemployee%5D/route-56da7df76f2c6417.js",revision:"56da7df76f2c6417"},{url:"/_next/static/chunks/app/api/(admins)/employeeInfo/%5Bemployee%5D/route-87ae0eb857486a0e.js",revision:"87ae0eb857486a0e"},{url:"/_next/static/chunks/app/api/(admins)/employeeNameAndPhoto/%5Bemployee%5D/route-28b409cf922dc67a.js",revision:"28b409cf922dc67a"},{url:"/_next/static/chunks/app/api/(admins)/getAllActiveEmployeeName/route-f1484d38a4482538.js",revision:"f1484d38a4482538"},{url:"/_next/static/chunks/app/api/(admins)/getAllCostCodes/route-4a385bccde015a04.js",revision:"4a385bccde015a04"},{url:"/_next/static/chunks/app/api/(admins)/getAllCostCodesByJobSites/route-2aa98c922c915a75.js",revision:"2aa98c922c915a75"},{url:"/_next/static/chunks/app/api/(admins)/getAllCrews/route-57856ab5f28317d0.js",revision:"57856ab5f28317d0"},{url:"/_next/static/chunks/app/api/(admins)/getAllEmployees/route-c770905efbb4d90d.js",revision:"c770905efbb4d90d"},{url:"/_next/static/chunks/app/api/(admins)/getAllEquipment/route-62ff392d532e9cfc.js",revision:"62ff392d532e9cfc"},{url:"/_next/static/chunks/app/api/(admins)/getAllForms/route-8ba38a4fd49f1a6e.js",revision:"8ba38a4fd49f1a6e"},{url:"/_next/static/chunks/app/api/(admins)/getAllTimesheetInfo/route-9140a1c60b8c6558.js",revision:"9140a1c60b8c6558"},{url:"/_next/static/chunks/app/api/(admins)/getAllTimesheetsPending/route-005a4dddf9658945.js",revision:"005a4dddf9658945"},{url:"/_next/static/chunks/app/api/(admins)/getClientById/%5Bid%5D/route-73bd3f6900a9bf25.js",revision:"73bd3f6900a9bf25"},{url:"/_next/static/chunks/app/api/(admins)/getClientsDetails/route-8d2e55db6c7ae63e.js",revision:"8d2e55db6c7ae63e"},{url:"/_next/static/chunks/app/api/(admins)/getClientsSummary/route-a121452bcf65d6f3.js",revision:"a121452bcf65d6f3"},{url:"/_next/static/chunks/app/api/(admins)/getContacts/route-518674b8058a59cc.js",revision:"518674b8058a59cc"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeById/%5Bid%5D/route-733e5b7bdf046431.js",revision:"733e5b7bdf046431"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeDetails/route-56b6f4fadfdb78f3.js",revision:"56b6f4fadfdb78f3"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeSummary/route-434f4e4c55385d3c.js",revision:"434f4e4c55385d3c"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeTags/%5Bid%5D/route-dd27314944b67d0b.js",revision:"dd27314944b67d0b"},{url:"/_next/static/chunks/app/api/(admins)/getCrewByIdAdmin/%5Bid%5D/route-483e5bdd96d9b7a6.js",revision:"483e5bdd96d9b7a6"},{url:"/_next/static/chunks/app/api/(admins)/getDashboard/route-0937a804dc57329e.js",revision:"0937a804dc57329e"},{url:"/_next/static/chunks/app/api/(admins)/getEmployee/route-e59ed3976f190b55.js",revision:"e59ed3976f190b55"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentByEquipmentId/%5Bequipment%5D/route-01fe3331a0db35b7.js",revision:"01fe3331a0db35b7"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentDetails/route-a78f3125f0506850.js",revision:"a78f3125f0506850"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentSummary/route-68ab415330904d17.js",revision:"68ab415330904d17"},{url:"/_next/static/chunks/app/api/(admins)/getFormSubmissionsById/%5Bid%5D/route-08006c77c31b7082.js",revision:"08006c77c31b7082"},{url:"/_next/static/chunks/app/api/(admins)/getForms/%5Bid%5D/route-3beb67c26f077db6.js",revision:"3beb67c26f077db6"},{url:"/_next/static/chunks/app/api/(admins)/getJobsiteById/%5Bid%5D/route-b185ed3ca89217a3.js",revision:"b185ed3ca89217a3"},{url:"/_next/static/chunks/app/api/(admins)/getJobsiteSummary/route-17fd22f3a1c540ad.js",revision:"17fd22f3a1c540ad"},{url:"/_next/static/chunks/app/api/(admins)/getMechanicProjectSummary/route-6e584876c11632d0.js",revision:"6e584876c11632d0"},{url:"/_next/static/chunks/app/api/(admins)/getReport/route-d584aa68acdb9736.js",revision:"d584aa68acdb9736"},{url:"/_next/static/chunks/app/api/(admins)/getTagById/%5Bid%5D/route-4df927e89808ebe0.js",revision:"4df927e89808ebe0"},{url:"/_next/static/chunks/app/api/(admins)/getTagSummary/route-383f5b98942809ac.js",revision:"383f5b98942809ac"},{url:"/_next/static/chunks/app/api/(admins)/getTimesheetById/route-29c1f34b3694993a.js",revision:"29c1f34b3694993a"},{url:"/_next/static/chunks/app/api/(admins)/jobsiteManager/route-f3f77eb6bd43e4f9.js",revision:"f3f77eb6bd43e4f9"},{url:"/_next/static/chunks/app/api/(admins)/reports/tasco/route-74dcba5c104fac3b.js",revision:"74dcba5c104fac3b"},{url:"/_next/static/chunks/app/api/(admins)/reports/truckingReport/route-cb86636a10c194b7.js",revision:"cb86636a10c194b7"},{url:"/_next/static/chunks/app/api/(clock)/getEquipmentList/route-519804720da908d7.js",revision:"519804720da908d7"},{url:"/_next/static/chunks/app/api/(clock)/getLogs/route-a3c44689dc229dcd.js",revision:"a3c44689dc229dcd"},{url:"/_next/static/chunks/app/api/(clock)/getMyTeamsUsers/route-6f02a95895bd7580.js",revision:"6f02a95895bd7580"},{url:"/_next/static/chunks/app/api/(clock)/getPendingTeamTimesheets/%5BcrewMembers%5D/route-49a6fb587a68d661.js",revision:"49a6fb587a68d661"},{url:"/_next/static/chunks/app/api/(clock)/getRecentCostCodes/route-7b2a5a6bdd75cb56.js",revision:"7b2a5a6bdd75cb56"},{url:"/_next/static/chunks/app/api/(clock)/getTodaysTimesheets/route-0e3e4464df7bb4bd.js",revision:"0e3e4464df7bb4bd"},{url:"/_next/static/chunks/app/api/(clock)/getTruckData/route-11b15a94c89f3a17.js",revision:"11b15a94c89f3a17"},{url:"/_next/static/chunks/app/api/(equipment)/getAllEquipmentIdAndQrId/route-a86c0b52bf966423.js",revision:"a86c0b52bf966423"},{url:"/_next/static/chunks/app/api/(equipment)/getCheckedList/route-29cf46270b7c0b16.js",revision:"29cf46270b7c0b16"},{url:"/_next/static/chunks/app/api/(equipment)/getEqUserLogs/%5Bid%5D/route-5031e3c8665a3a8a.js",revision:"5031e3c8665a3a8a"},{url:"/_next/static/chunks/app/api/(equipment)/getEqUserLogs/route-9e1429a0f3ecf97a.js",revision:"9e1429a0f3ecf97a"},{url:"/_next/static/chunks/app/api/(equipment)/getEquipmentRefueledLogs/%5Bid%5D/route-154e76b27348348e.js",revision:"154e76b27348348e"},{url:"/_next/static/chunks/app/api/(equipment)/getRecentJobDetails/route-593c8a1b40146ccc.js",revision:"593c8a1b40146ccc"},{url:"/_next/static/chunks/app/api/(forms)/employeeRequests/%5Bfilter%5D/route-79b1b4688fe82749.js",revision:"79b1b4688fe82749"},{url:"/_next/static/chunks/app/api/(forms)/form/%5Bid%5D/route-5cc9fdcfe1038764.js",revision:"5cc9fdcfe1038764"},{url:"/_next/static/chunks/app/api/(forms)/formDraft/%5Bid%5D/route-38d0f679bc982e5b.js",revision:"38d0f679bc982e5b"},{url:"/_next/static/chunks/app/api/(forms)/formDrafts/route-09288ae8c73e0c79.js",revision:"09288ae8c73e0c79"},{url:"/_next/static/chunks/app/api/(forms)/formSubmission/%5Bid%5D/route-c5c102d4a467d91e.js",revision:"c5c102d4a467d91e"},{url:"/_next/static/chunks/app/api/(forms)/formSubmissions/%5Bstatus%5D/route-7a0273cf39dae2ba.js",revision:"7a0273cf39dae2ba"},{url:"/_next/static/chunks/app/api/(forms)/forms/route-28e2a731008e9bc7.js",revision:"28e2a731008e9bc7"},{url:"/_next/static/chunks/app/api/(forms)/getDocumentById/%5Bid%5D/route-4205af3ca921846b.js",revision:"4205af3ca921846b"},{url:"/_next/static/chunks/app/api/(forms)/getDocumentTags/route-2ab76c44117504b8.js",revision:"2ab76c44117504b8"},{url:"/_next/static/chunks/app/api/(forms)/getDocuments/route-871245f433d876c3.js",revision:"871245f433d876c3"},{url:"/_next/static/chunks/app/api/(forms)/getEmployees/route-351b5ca277a5cc1a.js",revision:"351b5ca277a5cc1a"},{url:"/_next/static/chunks/app/api/(forms)/getEqForDocs/route-739a19704eb7f5a5.js",revision:"739a19704eb7f5a5"},{url:"/_next/static/chunks/app/api/(forms)/managerFormApproval/%5Bid%5D/route-5201e65a13478ba4.js",revision:"5201e65a13478ba4"},{url:"/_next/static/chunks/app/api/(forms)/teamSubmission/%5Bid%5D/route-bb8ab9a66aa355af.js",revision:"bb8ab9a66aa355af"},{url:"/_next/static/chunks/app/api/(general)/cookies/route-f7ea059705edfe5b.js",revision:"f7ea059705edfe5b"},{url:"/_next/static/chunks/app/api/(general)/getAllJobsites/route-5aedb815f8f934ab.js",revision:"5aedb815f8f934ab"},{url:"/_next/static/chunks/app/api/(general)/getAllTags/route-4899eaf2afc5cbb9.js",revision:"4899eaf2afc5cbb9"},{url:"/_next/static/chunks/app/api/(general)/getBannerData/route-253b4ba6c6587580.js",revision:"253b4ba6c6587580"},{url:"/_next/static/chunks/app/api/(general)/getCostCodes/route-32e9164c947e6c1b.js",revision:"32e9164c947e6c1b"},{url:"/_next/static/chunks/app/api/(general)/getCrewByCrewId/%5Bcrew%5D/route-dcf87156946b685e.js",revision:"dcf87156946b685e"},{url:"/_next/static/chunks/app/api/(general)/getEquipment/route-a6dda7b148dac447.js",revision:"a6dda7b148dac447"},{url:"/_next/static/chunks/app/api/(general)/getPayPeriodTimeSheets/route-ecf0ced3fc49b439.js",revision:"ecf0ced3fc49b439"},{url:"/_next/static/chunks/app/api/(general)/getRecentEquipment/route-820910284410c9dc.js",revision:"820910284410c9dc"},{url:"/_next/static/chunks/app/api/(general)/getRecentJobsites/route-8102539df32e6741.js",revision:"8102539df32e6741"},{url:"/_next/static/chunks/app/api/(general)/getRecentTimecard/route-7b01d2b3275d2e61.js",revision:"7b01d2b3275d2e61"},{url:"/_next/static/chunks/app/api/(general)/getRecentTimecardReturn/route-3bbf0b2bd9db103e.js",revision:"3bbf0b2bd9db103e"},{url:"/_next/static/chunks/app/api/(general)/getSettings/route-643152c22a75a21c.js",revision:"643152c22a75a21c"},{url:"/_next/static/chunks/app/api/(general)/getTimesheets/%5Bemployee%5D/route-977fee42de5e0bb1.js",revision:"977fee42de5e0bb1"},{url:"/_next/static/chunks/app/api/(general)/getTimesheets/route-a9c286c949ece94b.js",revision:"a9c286c949ece94b"},{url:"/_next/static/chunks/app/api/(general)/getUserSignature/route-bfc644e772bf5c5f.js",revision:"bfc644e772bf5c5f"},{url:"/_next/static/chunks/app/api/(mechanic)/getMaintenanceLogs/%5Bid%5D/route-3501e832c760ef0e.js",revision:"3501e832c760ef0e"},{url:"/_next/static/chunks/app/api/(mechanic)/getMaintenanceProjects/route-56fd3b1b9f7f257d.js",revision:"56fd3b1b9f7f257d"},{url:"/_next/static/chunks/app/api/(mechanic)/getReceivedInfo/%5Bid%5D/route-a18d16fae8d59d27.js",revision:"a18d16fae8d59d27"},{url:"/_next/static/chunks/app/api/(mechanic)/getRepairDetails/%5Bid%5D/route-ff9286d6db91936a.js",revision:"ff9286d6db91936a"},{url:"/_next/static/chunks/app/api/(myTeam)/getCrew/route-6189b264a6a17479.js",revision:"6189b264a6a17479"},{url:"/_next/static/chunks/app/api/(myTeam)/getCrewById/%5BcrewId%5D/route-c3d9fa8b97fa5c7b.js",revision:"c3d9fa8b97fa5c7b"},{url:"/_next/static/chunks/app/api/(myTeam)/getPendingTeamTimeSheets/route-adaf469f7759eefa.js",revision:"adaf469f7759eefa"},{url:"/_next/static/chunks/app/api/(myTeam)/getTeam/route-3cf21d9cc369f46c.js",revision:"3cf21d9cc369f46c"},{url:"/_next/static/chunks/app/api/(myTeam)/getUserInfo/%5Bid%5D/route-a8b3b9b339d77de8.js",revision:"a8b3b9b339d77de8"},{url:"/_next/static/chunks/app/api/(qrGenerator)/getEquipmentbyQrId/%5BqrId%5D/route-f3750198a0eff189.js",revision:"f3750198a0eff189"},{url:"/_next/static/chunks/app/api/(qrGenerator)/getJobsites/route-1260a25b244737a9.js",revision:"1260a25b244737a9"},{url:"/_next/static/chunks/app/api/(reviewWithDevun)/getSignature/%5Bid%5D/route-0d74e81c1ffdb230.js",revision:"0d74e81c1ffdb230"},{url:"/_next/static/chunks/app/api/(reviewWithDevun)/getTimesheetById/%5Bemployee%5D/route-f0394b3155ea47a1.js",revision:"f0394b3155ea47a1"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/comment/%5BtimeSheetId%5D/route-78e61373d1cab7ce.js",revision:"78e61373d1cab7ce"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/loadCount/%5BtimeSheetId%5D/route-a862ff6accf96c4b.js",revision:"a862ff6accf96c4b"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/refueledLogs/%5BtimeSheetId%5D/route-6896cc408bfead82.js",revision:"6896cc408bfead82"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/tascoId/route-6b984e161dc8640c.js",revision:"6b984e161dc8640c"},{url:"/_next/static/chunks/app/api/(timeSheets)/getComment/route-b96bc0053c8285ca.js",revision:"b96bc0053c8285ca"},{url:"/_next/static/chunks/app/api/(timeSheets)/getMaterialTypes/route-a6e81d8e0f25ce1f.js",revision:"a6e81d8e0f25ce1f"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/endingMileage/%5BtimeSheetId%5D/route-6400ee7d114ab76e.js",revision:"6400ee7d114ab76e"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/equipmentHauled/%5BtimeSheetId%5D/route-dae715faacbdc170.js",revision:"dae715faacbdc170"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/material/%5BtimeSheetId%5D/route-e14f706ef38876ad.js",revision:"e14f706ef38876ad"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/notes/%5BtimeSheetId%5D/route-a8df566facdb87a5.js",revision:"a8df566facdb87a5"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/refueledLogs/%5BtimeSheetId%5D/route-d707528893612245.js",revision:"d707528893612245"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/startingMileage/%5BtimeSheetId%5D/route-73979c38bc3c694a.js",revision:"73979c38bc3c694a"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/stateMileage/%5BtimeSheetId%5D/route-58292ddbc71f8c48.js",revision:"58292ddbc71f8c48"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/truckingId/route-96417f5ccf46d783.js",revision:"96417f5ccf46d783"},{url:"/_next/static/chunks/app/api/(user)/getUserImage/route-ead4f465a0cffdf7.js",revision:"ead4f465a0cffdf7"},{url:"/_next/static/chunks/app/api/auth/%5B...nextauth%5D/route-7ccbfd55f62725ba.js",revision:"7ccbfd55f62725ba"},{url:"/_next/static/chunks/app/api/clockoutDetails/route-adc8b0b486305055.js",revision:"adc8b0b486305055"},{url:"/_next/static/chunks/app/api/equipment/%5BequipmentId%5D/lastMileage/route-886eedfd2aa2cce3.js",revision:"886eedfd2aa2cce3"},{url:"/_next/static/chunks/app/api/getTimesheetDetailsManager/%5Bid%5D/route-5465e55fc785b46e.js",revision:"5465e55fc785b46e"},{url:"/_next/static/chunks/app/api/getTimesheetsByDate/route-f5d77af050d64c1e.js",revision:"f5d77af050d64c1e"},{url:"/_next/static/chunks/app/api/getTimesheetsByDateNew/route-f7724fd5ba534f70.js",revision:"f7724fd5ba534f70"},{url:"/_next/static/chunks/app/api/getTruckingLogs/equipmentId/%5BtimeSheetId%5D/route-92e9076b3f9b6302.js",revision:"92e9076b3f9b6302"},{url:"/_next/static/chunks/app/api/sentry-example-api/route-28b9c2ee47fa956d.js",revision:"28b9c2ee47fa956d"},{url:"/_next/static/chunks/app/api/truckingLogDetails/%5BtruckingLogId%5D/route-63b60b14edffe07b.js",revision:"63b60b14edffe07b"},{url:"/_next/static/chunks/app/api/user-permissions/route-3b37bd7da01b67a2.js",revision:"3b37bd7da01b67a2"},{url:"/_next/static/chunks/app/error-8ed5a2d151f85834.js",revision:"8ed5a2d151f85834"},{url:"/_next/static/chunks/app/global-error-293467dba0cb08e6.js",revision:"293467dba0cb08e6"},{url:"/_next/static/chunks/app/icons/page-c8cf36ee1169cb53.js",revision:"c8cf36ee1169cb53"},{url:"/_next/static/chunks/app/layout-9c800248433d75d3.js",revision:"9c800248433d75d3"},{url:"/_next/static/chunks/app/not-found-11295bbc0cdb94b7.js",revision:"11295bbc0cdb94b7"},{url:"/_next/static/chunks/app/page-545d0c878cb6aaa6.js",revision:"545d0c878cb6aaa6"},{url:"/_next/static/chunks/app/sentry-example-page/page-fc2572366eab84f9.js",revision:"fc2572366eab84f9"},{url:"/_next/static/chunks/framework-5871c7b6f4022a8e.js",revision:"5871c7b6f4022a8e"},{url:"/_next/static/chunks/main-app-3f7a4d6b51031eab.js",revision:"3f7a4d6b51031eab"},{url:"/_next/static/chunks/main-d1796a2f9f20c955.js",revision:"d1796a2f9f20c955"},{url:"/_next/static/chunks/pages/_app-5376b67503103dac.js",revision:"5376b67503103dac"},{url:"/_next/static/chunks/pages/_error-e94a0482ccaf0c05.js",revision:"e94a0482ccaf0c05"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-4911d3486e27635d.js",revision:"4911d3486e27635d"},{url:"/_next/static/css/0265c5315f14412e.css",revision:"0265c5315f14412e"},{url:"/_next/static/css/63c7182515913e2f.css",revision:"63c7182515913e2f"},{url:"/_next/static/css/ae3f615d9a81d81a.css",revision:"ae3f615d9a81d81a"},{url:"/_next/static/css/b5fe2f5a95132839.css",revision:"b5fe2f5a95132839"},{url:"/_next/static/css/ccce47fdcedd91d2.css",revision:"ccce47fdcedd91d2"},{url:"/_next/static/media/30e6a981a0efeb9b-s.woff2",revision:"5cd14db2bb47ab5db74376b77c909759"},{url:"/_next/static/media/5df40a3eb8df1a4c-s.p.woff2",revision:"2f46a0d75dfff5f623a08402583e7ee9"},{url:"/_next/static/media/62d466ce3199db61-s.p.woff2",revision:"3a28d888908be7f560bf59faa1c796b2"},{url:"/_next/static/media/92c2336771404627-s.woff2",revision:"e456d9e276ba68ad76f65962e88fab81"},{url:"/_next/static/media/ajax-loader.0b80f665.gif",revision:"0b80f665"},{url:"/_next/static/media/d934919622a522a7-s.woff2",revision:"d3322001406183e84b58b99afd278b84"},{url:"/_next/static/media/slick.25572f22.eot",revision:"25572f22"},{url:"/_next/static/media/slick.653a4cbb.woff",revision:"653a4cbb"},{url:"/_next/static/media/slick.6aa1ee46.ttf",revision:"6aa1ee46"},{url:"/_next/static/media/slick.f895cfdf.svg",revision:"f895cfdf"},{url:"/_next/static/nI36YKBSPGhE6_oCJsWS2/_buildManifest.js",revision:"0e206a604e8ff252314efab19db4eacf"},{url:"/_next/static/nI36YKBSPGhE6_oCJsWS2/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/admin-white.svg",revision:"f60bbe82328c01c11b8e0b2970c889e6"},{url:"/admin.svg",revision:"0d220493364872b69d6f6cde6e385d2c"},{url:"/android/android-launchericon-144-144.png",revision:"554cb47f9b447d7c049e0af337f30598"},{url:"/android/android-launchericon-192-192.png",revision:"c7acce3a1ca251dc273817cd5462a666"},{url:"/android/android-launchericon-48-48.png",revision:"b3e487bbda706b0a7a31bd383fcb5bb9"},{url:"/android/android-launchericon-512-512.png",revision:"daaca9c8424139dcfdeed3181b0a7cbd"},{url:"/android/android-launchericon-72-72.png",revision:"0544816fbb562d0825feb6e9ff73edff"},{url:"/android/android-launchericon-96-96.png",revision:"29c67668a313466cbd66948974b9c656"},{url:"/arrowBack.svg",revision:"17d544aac2962f3f0b83d7dd8bf8d5e7"},{url:"/arrowDown.svg",revision:"bde6e8ea1fa48902e773fc0f34d8ad2c"},{url:"/arrowDownThin.svg",revision:"c8dc31e8dcdca0b4feb827ad7b28c336"},{url:"/arrowLeft.svg",revision:"548bd36e974c05627ee6042a0d17e6c6"},{url:"/arrowLeftSymbol.svg",revision:"4cfdc72b6b5f2cf11ea191dc60c5c4b4"},{url:"/arrowLeftThin.svg",revision:"c4bbdff0ab65d617688561fee993f1c1"},{url:"/arrowRight.svg",revision:"36ae3b90062b7432459cad98be741a80"},{url:"/arrowRightSymbol.svg",revision:"e771082a8a75d50f59df162c67d26091"},{url:"/arrowRightThin.svg",revision:"73a5a9d7e4bf4951d11c4e268faeee85"},{url:"/arrowUp.svg",revision:"3e3a91581069f2b5dc59c7c465255263"},{url:"/arrowUpThin.svg",revision:"ef72218a70f19849149176593f9a8cee"},{url:"/biometrics.svg",revision:"f42bc2055fc95544b3641a2819a084d7"},{url:"/broken.svg",revision:"c62a555a912a5887033a4661236ae31d"},{url:"/calendar.svg",revision:"d6aa1ba02d5a6253fcec8da68242da2a"},{url:"/camera.svg",revision:"4657cc3ac5e2a543f0829e6100446591"},{url:"/cameraFilled.svg",revision:"bcdbce7deaeb9a1d37058d3befea92fc"},{url:"/checkbox.svg",revision:"c2c40521ced6fc4ba77a3865d9dc345f"},{url:"/clock.svg",revision:"eff35cd79029bdd0eeebc39b052f0fbf"},{url:"/clockBreak.svg",revision:"0f0d1da56fbbed48b97b5dbba7d34e65"},{url:"/clockIn.svg",revision:"02e58b9c6f4a129a7026400944bd15b7"},{url:"/clockOut.svg",revision:"a8b9f471daf782532b4006c2cb20b89a"},{url:"/comment.svg",revision:"7bb5715a71c021149b5c3b811aac572b"},{url:"/comments.svg",revision:"29b1a7eb27c708878a454c2eeaad616d"},{url:"/condense-white.svg",revision:"6e9075aa9d7fcb1c242b78610ea9330a"},{url:"/condense.svg",revision:"bc4d6b17d44724df96d6d4a3016072da"},{url:"/drag.svg",revision:"da36f043748f4e8af0e0f5ee17e289a2"},{url:"/dragDots.svg",revision:"c9d25bda626917ea72ca49977f11a016"},{url:"/dragFormBuilder.svg",revision:"bfdedbf1bc6d64afb3642848a542f55d"},{url:"/endDay.svg",revision:"78010192db1eaa29e95ac4d2d9f3705e"},{url:"/equipment-white.svg",revision:"0760c1e58029025d443832cdc33d7402"},{url:"/equipment.svg",revision:"eeb87295d51e91e2adb6483c18bc2452"},{url:"/eraser.svg",revision:"7ed7c543a0ff5530d3c586b368c51315"},{url:"/export-white.svg",revision:"4793d42ec50e11a8586bbc2a5a2f59c8"},{url:"/export.svg",revision:"09f9524501c9e871487735c2b09fcc43"},{url:"/eye-blue.svg",revision:"a851e7f5c035279390dfa1eaa3387e0d"},{url:"/eye.svg",revision:"b7243fcb645ec88a90d9b66fdc2abca3"},{url:"/eyeSlash.svg",revision:"3bb6de382b51b3a333e146bcf8a5d4f5"},{url:"/favicon_io/android-chrome-192x192.png",revision:"07281900d7ffa5690cc984da6e613805"},{url:"/favicon_io/android-chrome-512x512.png",revision:"ee520bf09fd4e101b8952887694905fc"},{url:"/favicon_io/apple-touch-icon.png",revision:"6f682a595255e6aab2bf0f2daa78403a"},{url:"/favicon_io/favicon-16x16.png",revision:"2d00862eea3e0a26607d3d0ce24d1747"},{url:"/favicon_io/favicon-32x32.png",revision:"84f7a4563c6fb8cd905f2dbedcffede8"},{url:"/fileClosed.svg",revision:"bf5a8aeebd62c1736f9e0062a3a271a0"},{url:"/fileOpen.svg",revision:"9ce1ceadae4b4abba237029e0a410d8f"},{url:"/filterDials.svg",revision:"13341b5339e076be7b361c2e09e4976c"},{url:"/filterFunnel.svg",revision:"743604a4e2e50abbb135746948880f65"},{url:"/form-white.svg",revision:"5f85274db63c253161739424aa2fab34"},{url:"/form.svg",revision:"5828ab7b08b01307092a48136ac39cb0"},{url:"/formApproval.svg",revision:"a53cba1fffa0cf55e4dbd86d0345b5e8"},{url:"/formDuplicate.svg",revision:"898bd36ec2c706978dae4e684575ec3e"},{url:"/formEdit.svg",revision:"017c59837c434df47dce4b457e71996e"},{url:"/formInspect-white.svg",revision:"8e48a64eb645a8b5bdc204706e8e6451"},{url:"/formInspect.svg",revision:"f2fcd14456afc03a583c568f4c4da3c8"},{url:"/formList-white.svg",revision:"e1ced4954a119850e3f06e0980062497"},{url:"/formList.svg",revision:"aef95c6f9019999cb9992dd2a21b3975"},{url:"/formSave.svg",revision:"e504c19b5bbbe70b0d2c83c9a73927b8"},{url:"/formSent.svg",revision:"eca064a2079c76fd61ca20528716ceee"},{url:"/formUndo.svg",revision:"91f221fce7b16f414253f9f1873902b2"},{url:"/hauling.svg",revision:"57a718f3ba6755d80106a01a21063d79"},{url:"/haulingFilled.svg",revision:"ebe07b7f9ca47cc45f7fd6b48e6d02ca"},{url:"/header.svg",revision:"479adabc8b70a52625854fba7e433548"},{url:"/home-white.svg",revision:"50f6d70fd4e0a491044340d69138b394"},{url:"/home.svg",revision:"c78e854f7b56bb952a0a22fc1980bc7d"},{url:"/icon-192X192.png",revision:"47aa596be976071dfff9c1e5455ecc80"},{url:"/icon-384x384.png",revision:"99d1987d20b874836d950b95586968ae"},{url:"/icon-512x512.png",revision:"c16feee422d814c946012979c0d02f00"},{url:"/inbox-white.svg",revision:"dfb451099519b395ed3d9b15daad5708"},{url:"/inbox.svg",revision:"03332320e762076df175dde846d62022"},{url:"/inboxFilled.svg",revision:"8b5e6d06de80a9927ae85007277b98a9"},{url:"/information.svg",revision:"3e9a1c81c8fd396fdf6849975d62dbb5"},{url:"/injury.svg",revision:"18121c3e1dc867f5d0a332a7133057f4"},{url:"/ios/100.png",revision:"5222167951b54f250720976e5798bd81"},{url:"/ios/1024.png",revision:"4046bc171d373187ec11f28c06ceb334"},{url:"/ios/114.png",revision:"2aec0491cbf7eb22d34f3079cf2eff54"},{url:"/ios/120.png",revision:"52ac3d673f239d4c822b20c337c73db2"},{url:"/ios/128.png",revision:"bc9dd541b1a0aaff432df4f5b0fb88b4"},{url:"/ios/144.png",revision:"554cb47f9b447d7c049e0af337f30598"},{url:"/ios/152.png",revision:"3085835ae98a412b5f8bb41b3d7fccae"},{url:"/ios/16.png",revision:"876eb7a741ae46958961bf0ed6c30ba0"},{url:"/ios/167.png",revision:"a7230785a2891b4659792d7d6cab6ffb"},{url:"/ios/180.png",revision:"5cdb37c8967955e3a4fae232138f3c03"},{url:"/ios/192.png",revision:"c7acce3a1ca251dc273817cd5462a666"},{url:"/ios/20.png",revision:"8c0f62f1ddd74accb89e318fe432a70f"},{url:"/ios/256.png",revision:"19665fd60a705b586e5cc5c93eddc1f2"},{url:"/ios/29.png",revision:"536c6f0fcf359bba5e0b45c9bb4fc260"},{url:"/ios/32.png",revision:"55a60f12c21ea3fffa61d8f7a34cfc33"},{url:"/ios/40.png",revision:"a38e8ea2895f2cfe54be4ce7465e7952"},{url:"/ios/50.png",revision:"db849275edb671b9ea7c45ef6f4fc489"},{url:"/ios/512.png",revision:"daaca9c8424139dcfdeed3181b0a7cbd"},{url:"/ios/57.png",revision:"8511d6a6a25878417899d2041de166ff"},{url:"/ios/58.png",revision:"001948872d9d885d20140b7aa9cc6ff8"},{url:"/ios/60.png",revision:"80209b85fb0be10562bcee31908d2a95"},{url:"/ios/64.png",revision:"066fa2094719e23427225f5517889ed7"},{url:"/ios/72.png",revision:"0544816fbb562d0825feb6e9ff73edff"},{url:"/ios/76.png",revision:"0a1c5da7e5a9673345c44f7e1c87f0d4"},{url:"/ios/80.png",revision:"91b74cbd7d0f0119a164a0750462149e"},{url:"/ios/87.png",revision:"74c6a6624174372f1fc483515c820c2f"},{url:"/jobsite-white.svg",revision:"1d991bcf439a2538ed78c709c708954b"},{url:"/jobsite.svg",revision:"f3909b4ff16c4f25e211fb0dc7556050"},{url:"/key.svg",revision:"f39b7da7fb64da838435e228ca3113ed"},{url:"/language.svg",revision:"e4615572d416d5df80c463dc53d5e0b7"},{url:"/layout.svg",revision:"a3b50fb72e8fab8feac60ccbd864e31a"},{url:"/logo.svg",revision:"8e78aff787124cc6567c2ccfd520099d"},{url:"/logo_JPG.jpg",revision:"e84c38cbd142e140688beacd59e6a0cf"},{url:"/manifest.json",revision:"05d4e6a0d9c7c7bebec473d08ea582b2"},{url:"/mechanic.svg",revision:"4f39d9d1ff5c68a81b93c7f6611d8e9d"},{url:"/message.svg",revision:"c55f300ce914f4b28036e30580ee104c"},{url:"/mileage.svg",revision:"11e1eb1d01c9be54e3c8394e198db717"},{url:"/minus.svg",revision:"d2ec83afd3490cac9da747386df54184"},{url:"/moreOptions.svg",revision:"d3019cfec5d1102ec73147f1dd5e5faf"},{url:"/moreOptionsCircle.svg",revision:"02d86b2a2acd7707b4cef84da6ede884"},{url:"/number.svg",revision:"6f90bfd3f431fe8602537fe47b9a109b"},{url:"/plus-stroke-white.svg",revision:"f23aeed115ad4273e185d3e73a314ba7"},{url:"/plus-white.svg",revision:"6adf4b80867ce618d5ca70765c2e429d"},{url:"/plus.svg",revision:"15fadbc574e693de5d7b486f4e13b1d5"},{url:"/policies.svg",revision:"25de43083a5e24cdf83a7ba62d6c28c4"},{url:"/priorityDelay.svg",revision:"5286ef619f78a44968b600b83bba55a0"},{url:"/priorityHigh.svg",revision:"fa44198aecf39d4748914d7f3b4a0d03"},{url:"/priorityLow.svg",revision:"11efcad98767f7c049d90ece2f425f1a"},{url:"/priorityMedium.svg",revision:"219c8a19b23f94db24162d544e1605ec"},{url:"/priorityPending.svg",revision:"45d3ef273daf2b4daadcb0b6d0064789"},{url:"/priorityToday.svg",revision:"91fec1dfbc8a1268488c12a9ef9fa4ac"},{url:"/profileEmpty.svg",revision:"66491e65dd6114a9acf6990181fc44d9"},{url:"/profileFilled.svg",revision:"ff2e851ef303666374501b541725d34c"},{url:"/qrCode-white.svg",revision:"0bcc19567973489fa8a15e759b729d5b"},{url:"/qrCode.svg",revision:"a97e34eddfd5bae153771e734fd96e4c"},{url:"/radio.svg",revision:"17ecc9b4e325da65fc947c71d4bc1ade"},{url:"/refuel.svg",revision:"07fcabd768272b84f61007c4e130b8fc"},{url:"/refuelFilled.svg",revision:"d75e2ee23202812499b36b050fff841c"},{url:"/searchLeft.svg",revision:"7a4c52a1de03e0fbe1f2f949fed97367"},{url:"/searchRight.svg",revision:"9ac605dce8646197fa3ae4c90d93f8c9"},{url:"/settingsFilled.svg",revision:"4b9876c68beb7435b569e5a5a0252501"},{url:"/shiftScanLogo.svg",revision:"f6cdbf4c760167995264751f68ebb2af"},{url:"/shiftscanlogoHorizontal.svg",revision:"f6dac5483d88c1cdebc5f92190cc1d2f"},{url:"/spinner.svg",revision:"7eedb8067da00e29baf517b3cbfffc43"},{url:"/star.svg",revision:"744181e34e34c3385a845fa37cf51454"},{url:"/starFilled.svg",revision:"5ad7cd14d1deca592924c230b5a39218"},{url:"/state.svg",revision:"83c029784c461361996615f83c3850ef"},{url:"/stateFilled.svg",revision:"183c533850d5a2540b71b2c44bef9d11"},{url:"/statusApproved.svg",revision:"0e767f6c9f2c8f20480d3120d6b6e198"},{url:"/statusApprovedFilled.svg",revision:"6fc76b4c67d20fa9ffbdb451a30e8071"},{url:"/statusDenied.svg",revision:"7b1167c6857f8169396ae88ffbce95e6"},{url:"/statusDeniedFilled.svg",revision:"52740140811df8fc4cd87604621c9960"},{url:"/statusOffline.svg",revision:"d1931f81ba9fe9342e78e181c2216aaa"},{url:"/statusOngoing-white.svg",revision:"d07cb3a9608fbfa3e64486e49eb74365"},{url:"/statusOngoing.svg",revision:"7eedb8067da00e29baf517b3cbfffc43"},{url:"/statusOngoingFilled.svg",revision:"5e18aea67f178cddbdc7e2e449ea6e3a"},{url:"/statusOnline.svg",revision:"ebed6ae0692535bcfee10ada59c779c3"},{url:"/statusUnfinished.svg",revision:"038f7353e12e76758c9e214925a07a0c"},{url:"/swe-worker-5c72df51bb1f6ee0.js",revision:"447aee4a2c24fefe7917563f0255d688"},{url:"/tasco.svg",revision:"1280d07dd40b7604d022f71c3728ccfb"},{url:"/team-white.svg",revision:"0142ac2317a2b44b141c270916ab32ca"},{url:"/team.svg",revision:"203a48a7519f73fe7f24fff1cae20204"},{url:"/timecards.svg",revision:"7771f6f39b0e3d775e88c6d542d2d68b"},{url:"/tinyCheckMark-white.svg",revision:"a6d6eadc3dbfaafbf1e8257346eee02c"},{url:"/title.svg",revision:"15b1291ddf60f0ceb36ce46377fa6dd3"},{url:"/trash-red.svg",revision:"a42dfceb5b388b58d0c41c13ee0c17f7"},{url:"/trash.svg",revision:"229905df8ec07dd400f7a23deb582f0c"},{url:"/trucking.svg",revision:"ab3794a1a3073bc2926a1978a038300c"},{url:"/windows11/LargeTile.scale-100.png",revision:"e3ae3637973ea93ff81e73ee37effe0c"},{url:"/windows11/LargeTile.scale-125.png",revision:"6e4ec698528b71f25efd6ff2dd68e89f"},{url:"/windows11/LargeTile.scale-150.png",revision:"b636fa1812d337a4ca1224b433a3d855"},{url:"/windows11/LargeTile.scale-200.png",revision:"d18a045105d03ba23b624c109708f7bb"},{url:"/windows11/LargeTile.scale-400.png",revision:"d4176f9ebaacc0bbfebf56fd56e253ce"},{url:"/windows11/SmallTile.scale-100.png",revision:"bce1445e4fb91c8c0842780a4026522e"},{url:"/windows11/SmallTile.scale-125.png",revision:"9415b7ff287961c074dee1834b42aad9"},{url:"/windows11/SmallTile.scale-150.png",revision:"87d19729c39b8d355a62d180d91bb31b"},{url:"/windows11/SmallTile.scale-200.png",revision:"7c5db43e6077ca3c958c21c85bec6d25"},{url:"/windows11/SmallTile.scale-400.png",revision:"1913d6ed56bc5b31dc60858acbcace6f"},{url:"/windows11/SplashScreen.scale-100.png",revision:"a8215f24409e6231d1dc00a0907ed4a8"},{url:"/windows11/SplashScreen.scale-125.png",revision:"097ac942f0459d3481fbd62c0f97e299"},{url:"/windows11/SplashScreen.scale-150.png",revision:"201fc53647b740af27d9b004cafeaaf8"},{url:"/windows11/SplashScreen.scale-200.png",revision:"1b1081f94203f3823ed2dd1f8dabb29a"},{url:"/windows11/SplashScreen.scale-400.png",revision:"0d6ee6f415a7d2101792e2b8312af059"},{url:"/windows11/Square150x150Logo.scale-100.png",revision:"7b6855c0928dad5db20403341b059afc"},{url:"/windows11/Square150x150Logo.scale-125.png",revision:"dd8460f97de8d2cf5826468f8f8153db"},{url:"/windows11/Square150x150Logo.scale-150.png",revision:"8bfd4727bbeb856bd1af17e0cdef44ea"},{url:"/windows11/Square150x150Logo.scale-200.png",revision:"c95f1e579117442bcf6a388818260303"},{url:"/windows11/Square150x150Logo.scale-400.png",revision:"a20add32148ad6bfcc0d435ac4986eb5"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/Square44x44Logo.scale-100.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.scale-125.png",revision:"95498060191b331348eb8f5b5ab6dfe7"},{url:"/windows11/Square44x44Logo.scale-150.png",revision:"aced07a7a4627cdad487ed0a46faeebb"},{url:"/windows11/Square44x44Logo.scale-200.png",revision:"db11fa1c0b22e2344ecf9dea9af68a2c"},{url:"/windows11/Square44x44Logo.scale-400.png",revision:"5f1b23fb222c880bf3c533643f43b4ca"},{url:"/windows11/Square44x44Logo.targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/StoreLogo.scale-100.png",revision:"db849275edb671b9ea7c45ef6f4fc489"},{url:"/windows11/StoreLogo.scale-125.png",revision:"728e6527e2d77371e3007caf007b116f"},{url:"/windows11/StoreLogo.scale-150.png",revision:"9224f9a45257c2103e10aa9026af46d1"},{url:"/windows11/StoreLogo.scale-200.png",revision:"5222167951b54f250720976e5798bd81"},{url:"/windows11/StoreLogo.scale-400.png",revision:"6562c92cea67a6232c5727e464a9dd83"},{url:"/windows11/Wide310x150Logo.scale-100.png",revision:"7fc14247c4af58216f16d5f4f98fc46b"},{url:"/windows11/Wide310x150Logo.scale-125.png",revision:"a0c8d4ccacb2bc65438458df95ab3d60"},{url:"/windows11/Wide310x150Logo.scale-150.png",revision:"34161357c0cdafe2651b8d197c019798"},{url:"/windows11/Wide310x150Logo.scale-200.png",revision:"a8215f24409e6231d1dc00a0907ed4a8"},{url:"/windows11/Wide310x150Logo.scale-400.png",revision:"1b1081f94203f3823ed2dd1f8dabb29a"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https?.*/,new e.NetworkFirst({cacheName:"offlineCache",plugins:[new e.ExpirationPlugin({maxEntries:200,maxAgeSeconds:86400})]}),"GET"),self.__WB_DISABLE_DEV_LOGS=!0}));
+if(!self.define){let e,a={};const i=(i,s)=>(i=new URL(i+".js",s).href,a[i]||new Promise((a=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=a,document.head.appendChild(e)}else e=i,importScripts(i),a()})).then((()=>{let e=a[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e})));self.define=(s,c)=>{const r=e||("document"in self?document.currentScript.src:"")||location.href;if(a[r])return;let n={};const t=e=>i(e,r),d={module:{uri:r},exports:n,require:t};a[r]=Promise.all(s.map((e=>d[e]||t(e)))).then((e=>(c(...e),n)))}}define(["./workbox-31da5539"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/Settings.svg",revision:"ebe8b6a8bff9bd2fe8669ad036e8edf8"},{url:"/_next/static/chunks/1183-1edb584d8d091e42.js",revision:"1edb584d8d091e42"},{url:"/_next/static/chunks/1401-0de5bc94198b3c9a.js",revision:"0de5bc94198b3c9a"},{url:"/_next/static/chunks/1692-cc16fa14e0e5bbf3.js",revision:"cc16fa14e0e5bbf3"},{url:"/_next/static/chunks/1791-a4cd8ef707d3c1ff.js",revision:"a4cd8ef707d3c1ff"},{url:"/_next/static/chunks/2019-244f4105c4691c0c.js",revision:"244f4105c4691c0c"},{url:"/_next/static/chunks/2057-668deaa0eb77c6a4.js",revision:"668deaa0eb77c6a4"},{url:"/_next/static/chunks/2170a4aa-8c96009f60e20411.js",revision:"8c96009f60e20411"},{url:"/_next/static/chunks/2232-3f04a14cda4cd5a1.js",revision:"3f04a14cda4cd5a1"},{url:"/_next/static/chunks/2403.ca107014ddba66de.js",revision:"ca107014ddba66de"},{url:"/_next/static/chunks/2441-9ad2299f689922c4.js",revision:"9ad2299f689922c4"},{url:"/_next/static/chunks/2927-0615eaa2f61c8eef.js",revision:"0615eaa2f61c8eef"},{url:"/_next/static/chunks/3008-31aa8928ac008967.js",revision:"31aa8928ac008967"},{url:"/_next/static/chunks/3031-e0cbd25419c08505.js",revision:"e0cbd25419c08505"},{url:"/_next/static/chunks/3068-f2d5b17c55eef44e.js",revision:"f2d5b17c55eef44e"},{url:"/_next/static/chunks/3328-1e73249e263f1cb2.js",revision:"1e73249e263f1cb2"},{url:"/_next/static/chunks/3343-cddd99d48f629228.js",revision:"cddd99d48f629228"},{url:"/_next/static/chunks/4095-8d931f3f0218645a.js",revision:"8d931f3f0218645a"},{url:"/_next/static/chunks/4162-f6605c3f153eb555.js",revision:"f6605c3f153eb555"},{url:"/_next/static/chunks/4387-3676ac23a2682f13.js",revision:"3676ac23a2682f13"},{url:"/_next/static/chunks/460-8f004685eb760a9f.js",revision:"8f004685eb760a9f"},{url:"/_next/static/chunks/4610-9b372e525bb28e3c.js",revision:"9b372e525bb28e3c"},{url:"/_next/static/chunks/4692-79463bd8f0d76a0e.js",revision:"79463bd8f0d76a0e"},{url:"/_next/static/chunks/4bd1b696-88e37c56edce0f59.js",revision:"88e37c56edce0f59"},{url:"/_next/static/chunks/5152-9e6b4080454ffa3c.js",revision:"9e6b4080454ffa3c"},{url:"/_next/static/chunks/5163-a9ef2a6bf3806903.js",revision:"a9ef2a6bf3806903"},{url:"/_next/static/chunks/5166-22f32cf0bc0e8fc2.js",revision:"22f32cf0bc0e8fc2"},{url:"/_next/static/chunks/52774a7f-9f526229934f791c.js",revision:"9f526229934f791c"},{url:"/_next/static/chunks/5493-002a49a6d796cca8.js",revision:"002a49a6d796cca8"},{url:"/_next/static/chunks/5751-70590a584bd12447.js",revision:"70590a584bd12447"},{url:"/_next/static/chunks/602-c9c2a774f75643ad.js",revision:"c9c2a774f75643ad"},{url:"/_next/static/chunks/6323-d1e244e417681874.js",revision:"d1e244e417681874"},{url:"/_next/static/chunks/6544-bcf8e533ec6b3434.js",revision:"bcf8e533ec6b3434"},{url:"/_next/static/chunks/6671-c6ce7faab2868fe3.js",revision:"c6ce7faab2868fe3"},{url:"/_next/static/chunks/6842-d6266f7a919f2655.js",revision:"d6266f7a919f2655"},{url:"/_next/static/chunks/6862-81cfc3fff123ab8a.js",revision:"81cfc3fff123ab8a"},{url:"/_next/static/chunks/6874-bf3fa41ffddf2fd7.js",revision:"bf3fa41ffddf2fd7"},{url:"/_next/static/chunks/7053-2689323dc7a6c8d9.js",revision:"2689323dc7a6c8d9"},{url:"/_next/static/chunks/707-3d319102f6468d5f.js",revision:"3d319102f6468d5f"},{url:"/_next/static/chunks/7082-c91642ee82e5320d.js",revision:"c91642ee82e5320d"},{url:"/_next/static/chunks/7111-32719b32e4f3138d.js",revision:"32719b32e4f3138d"},{url:"/_next/static/chunks/7253-b6a7c4b8ba2a4ccc.js",revision:"b6a7c4b8ba2a4ccc"},{url:"/_next/static/chunks/7765-7d2d9a5f429672fa.js",revision:"7d2d9a5f429672fa"},{url:"/_next/static/chunks/789-86a3b92415cd819c.js",revision:"86a3b92415cd819c"},{url:"/_next/static/chunks/7914-e7d7b2a8c8cab0d5.js",revision:"e7d7b2a8c8cab0d5"},{url:"/_next/static/chunks/8007-9d7908d793cc20bb.js",revision:"9d7908d793cc20bb"},{url:"/_next/static/chunks/822-0edc64ee089dcd0d.js",revision:"0edc64ee089dcd0d"},{url:"/_next/static/chunks/8436.fe8b1513f8e56bfa.js",revision:"fe8b1513f8e56bfa"},{url:"/_next/static/chunks/9313-eae8d054304ecf16.js",revision:"eae8d054304ecf16"},{url:"/_next/static/chunks/9337-ed07232187c0fdae.js",revision:"ed07232187c0fdae"},{url:"/_next/static/chunks/9373-12ddcc70f6722621.js",revision:"12ddcc70f6722621"},{url:"/_next/static/chunks/949-b28f693ac97fe61e.js",revision:"b28f693ac97fe61e"},{url:"/_next/static/chunks/9648-fecb89b91c985f68.js",revision:"fecb89b91c985f68"},{url:"/_next/static/chunks/9763-3b7d58f5c287c919.js",revision:"3b7d58f5c287c919"},{url:"/_next/static/chunks/app/(routes)/admins/clients/page-3fa8a7055592d9d1.js",revision:"3fa8a7055592d9d1"},{url:"/_next/static/chunks/app/(routes)/admins/cost-codes/page-5b613be8b88a8ed4.js",revision:"5b613be8b88a8ed4"},{url:"/_next/static/chunks/app/(routes)/admins/equipment/page-c4c7aa3006934c70.js",revision:"c4c7aa3006934c70"},{url:"/_next/static/chunks/app/(routes)/admins/forms/%5Bid%5D/page-b1975f3c1114c1fd.js",revision:"b1975f3c1114c1fd"},{url:"/_next/static/chunks/app/(routes)/admins/forms/create/page-d7d1c50fe54db441.js",revision:"d7d1c50fe54db441"},{url:"/_next/static/chunks/app/(routes)/admins/forms/edit/%5Bid%5D/page-04cfc5424283b722.js",revision:"04cfc5424283b722"},{url:"/_next/static/chunks/app/(routes)/admins/forms/page-c4569757416362f7.js",revision:"c4569757416362f7"},{url:"/_next/static/chunks/app/(routes)/admins/jobsites/page-c9d2fff499f01793.js",revision:"c9d2fff499f01793"},{url:"/_next/static/chunks/app/(routes)/admins/layout-af7c0726e6b95d17.js",revision:"af7c0726e6b95d17"},{url:"/_next/static/chunks/app/(routes)/admins/page-111bcb9c8973e94e.js",revision:"111bcb9c8973e94e"},{url:"/_next/static/chunks/app/(routes)/admins/personnel/page-826711ddb5f9a33d.js",revision:"826711ddb5f9a33d"},{url:"/_next/static/chunks/app/(routes)/admins/reports/page-7548cd12dc68da90.js",revision:"7548cd12dc68da90"},{url:"/_next/static/chunks/app/(routes)/admins/timesheets/page-69c9ae94a4047824.js",revision:"69c9ae94a4047824"},{url:"/_next/static/chunks/app/(routes)/break/page-e81008d07e2321ea.js",revision:"e81008d07e2321ea"},{url:"/_next/static/chunks/app/(routes)/clock/page-f693f2e9ce928a45.js",revision:"f693f2e9ce928a45"},{url:"/_next/static/chunks/app/(routes)/dashboard/clock-out/(components)/log/page-8e60956c7cc32cd2.js",revision:"8e60956c7cc32cd2"},{url:"/_next/static/chunks/app/(routes)/dashboard/clock-out/page-be4435e6e9c4e7b7.js",revision:"be4435e6e9c4e7b7"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/%5Bid%5D/page-8e5b5fc236df03ee.js",revision:"8e5b5fc236df03ee"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/log-new/page-3f6c261ac7be890d.js",revision:"3f6c261ac7be890d"},{url:"/_next/static/chunks/app/(routes)/dashboard/equipment/page-78a72de45d1c6146.js",revision:"78a72de45d1c6146"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/edit-repair-details/%5Bid%5D/page-56d27e11be44efa6.js",revision:"56d27e11be44efa6"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/new-repair/page-492d857bb72f9d7d.js",revision:"492d857bb72f9d7d"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/page-bb69016e41d01914.js",revision:"bb69016e41d01914"},{url:"/_next/static/chunks/app/(routes)/dashboard/mechanic/projects/%5Bid%5D/page-a46164a20b3e6c25.js",revision:"a46164a20b3e6c25"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/%5Bid%5D/employee/%5BemployeeId%5D/page-120a6a015e972199.js",revision:"120a6a015e972199"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/%5Bid%5D/page-9d5a0f7593de0429.js",revision:"9d5a0f7593de0429"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/page-d56082030b4496ed.js",revision:"d56082030b4496ed"},{url:"/_next/static/chunks/app/(routes)/dashboard/myTeam/timecards/page-761165f83fc91545.js",revision:"761165f83fc91545"},{url:"/_next/static/chunks/app/(routes)/dashboard/page-4016bf3e76d83acb.js",revision:"4016bf3e76d83acb"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/add-equipment/page-1e648bfb9a300d32.js",revision:"1e648bfb9a300d32"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/add-jobsite/page-a89384fb40d47e71.js",revision:"a89384fb40d47e71"},{url:"/_next/static/chunks/app/(routes)/dashboard/qr-generator/page-a0594747f8a8bc58.js",revision:"a0594747f8a8bc58"},{url:"/_next/static/chunks/app/(routes)/dashboard/switch-jobs/page-b4e84d58fdc7476c.js",revision:"b4e84d58fdc7476c"},{url:"/_next/static/chunks/app/(routes)/dashboard/tasco/page-3bb97b3aea503c41.js",revision:"3bb97b3aea503c41"},{url:"/_next/static/chunks/app/(routes)/dashboard/truckingAssistant/page-fc0534ab53dede76.js",revision:"fc0534ab53dede76"},{url:"/_next/static/chunks/app/(routes)/hamburger/changePassword/page-9f6a991581c388a4.js",revision:"9f6a991581c388a4"},{url:"/_next/static/chunks/app/(routes)/hamburger/inbox/formSubmission/%5Bid%5D/page-60c440af1d4621cb.js",revision:"60c440af1d4621cb"},{url:"/_next/static/chunks/app/(routes)/hamburger/inbox/page-9dc20b54139982dd.js",revision:"9dc20b54139982dd"},{url:"/_next/static/chunks/app/(routes)/hamburger/profile/page-882d71f23505fdd6.js",revision:"882d71f23505fdd6"},{url:"/_next/static/chunks/app/(routes)/information/page-89d7583d61422b6d.js",revision:"89d7583d61422b6d"},{url:"/_next/static/chunks/app/(routes)/signin/forgot-password/page-b33d9bbb15f5cbfe.js",revision:"b33d9bbb15f5cbfe"},{url:"/_next/static/chunks/app/(routes)/signin/new-password/page-db8596a24fb7b927.js",revision:"db8596a24fb7b927"},{url:"/_next/static/chunks/app/(routes)/signin/page-776eda5d900bfbb8.js",revision:"776eda5d900bfbb8"},{url:"/_next/static/chunks/app/(routes)/signin/signup/page-b9664185c91a1fdf.js",revision:"b9664185c91a1fdf"},{url:"/_next/static/chunks/app/(routes)/timesheets/page-525cb415f3827045.js",revision:"525cb415f3827045"},{url:"/_next/static/chunks/app/_not-found/page-9c84c29c80583905.js",revision:"9c84c29c80583905"},{url:"/_next/static/chunks/app/api/(admins)/contactInfo/%5Bemployee%5D/route-56da7df76f2c6417.js",revision:"56da7df76f2c6417"},{url:"/_next/static/chunks/app/api/(admins)/employeeInfo/%5Bemployee%5D/route-87ae0eb857486a0e.js",revision:"87ae0eb857486a0e"},{url:"/_next/static/chunks/app/api/(admins)/employeeNameAndPhoto/%5Bemployee%5D/route-28b409cf922dc67a.js",revision:"28b409cf922dc67a"},{url:"/_next/static/chunks/app/api/(admins)/getAllActiveEmployeeName/route-f1484d38a4482538.js",revision:"f1484d38a4482538"},{url:"/_next/static/chunks/app/api/(admins)/getAllCostCodes/route-4a385bccde015a04.js",revision:"4a385bccde015a04"},{url:"/_next/static/chunks/app/api/(admins)/getAllCostCodesByJobSites/route-2aa98c922c915a75.js",revision:"2aa98c922c915a75"},{url:"/_next/static/chunks/app/api/(admins)/getAllCrews/route-57856ab5f28317d0.js",revision:"57856ab5f28317d0"},{url:"/_next/static/chunks/app/api/(admins)/getAllEmployees/route-c770905efbb4d90d.js",revision:"c770905efbb4d90d"},{url:"/_next/static/chunks/app/api/(admins)/getAllEquipment/route-62ff392d532e9cfc.js",revision:"62ff392d532e9cfc"},{url:"/_next/static/chunks/app/api/(admins)/getAllForms/route-8ba38a4fd49f1a6e.js",revision:"8ba38a4fd49f1a6e"},{url:"/_next/static/chunks/app/api/(admins)/getAllTimesheetInfo/route-9140a1c60b8c6558.js",revision:"9140a1c60b8c6558"},{url:"/_next/static/chunks/app/api/(admins)/getAllTimesheetsPending/route-005a4dddf9658945.js",revision:"005a4dddf9658945"},{url:"/_next/static/chunks/app/api/(admins)/getClientById/%5Bid%5D/route-73bd3f6900a9bf25.js",revision:"73bd3f6900a9bf25"},{url:"/_next/static/chunks/app/api/(admins)/getClientsDetails/route-8d2e55db6c7ae63e.js",revision:"8d2e55db6c7ae63e"},{url:"/_next/static/chunks/app/api/(admins)/getClientsSummary/route-a121452bcf65d6f3.js",revision:"a121452bcf65d6f3"},{url:"/_next/static/chunks/app/api/(admins)/getContacts/route-518674b8058a59cc.js",revision:"518674b8058a59cc"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeById/%5Bid%5D/route-733e5b7bdf046431.js",revision:"733e5b7bdf046431"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeDetails/route-56b6f4fadfdb78f3.js",revision:"56b6f4fadfdb78f3"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeSummary/route-434f4e4c55385d3c.js",revision:"434f4e4c55385d3c"},{url:"/_next/static/chunks/app/api/(admins)/getCostCodeTags/%5Bid%5D/route-dd27314944b67d0b.js",revision:"dd27314944b67d0b"},{url:"/_next/static/chunks/app/api/(admins)/getCrewByIdAdmin/%5Bid%5D/route-483e5bdd96d9b7a6.js",revision:"483e5bdd96d9b7a6"},{url:"/_next/static/chunks/app/api/(admins)/getDashboard/route-0937a804dc57329e.js",revision:"0937a804dc57329e"},{url:"/_next/static/chunks/app/api/(admins)/getEmployee/route-e59ed3976f190b55.js",revision:"e59ed3976f190b55"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentByEquipmentId/%5Bequipment%5D/route-01fe3331a0db35b7.js",revision:"01fe3331a0db35b7"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentDetails/route-a78f3125f0506850.js",revision:"a78f3125f0506850"},{url:"/_next/static/chunks/app/api/(admins)/getEquipmentSummary/route-68ab415330904d17.js",revision:"68ab415330904d17"},{url:"/_next/static/chunks/app/api/(admins)/getFormSubmissionsById/%5Bid%5D/route-08006c77c31b7082.js",revision:"08006c77c31b7082"},{url:"/_next/static/chunks/app/api/(admins)/getForms/%5Bid%5D/route-3beb67c26f077db6.js",revision:"3beb67c26f077db6"},{url:"/_next/static/chunks/app/api/(admins)/getJobsiteById/%5Bid%5D/route-b185ed3ca89217a3.js",revision:"b185ed3ca89217a3"},{url:"/_next/static/chunks/app/api/(admins)/getJobsiteSummary/route-17fd22f3a1c540ad.js",revision:"17fd22f3a1c540ad"},{url:"/_next/static/chunks/app/api/(admins)/getMechanicProjectSummary/route-6e584876c11632d0.js",revision:"6e584876c11632d0"},{url:"/_next/static/chunks/app/api/(admins)/getReport/route-d584aa68acdb9736.js",revision:"d584aa68acdb9736"},{url:"/_next/static/chunks/app/api/(admins)/getTagById/%5Bid%5D/route-4df927e89808ebe0.js",revision:"4df927e89808ebe0"},{url:"/_next/static/chunks/app/api/(admins)/getTagSummary/route-383f5b98942809ac.js",revision:"383f5b98942809ac"},{url:"/_next/static/chunks/app/api/(admins)/getTimesheetById/route-29c1f34b3694993a.js",revision:"29c1f34b3694993a"},{url:"/_next/static/chunks/app/api/(admins)/jobsiteManager/route-f3f77eb6bd43e4f9.js",revision:"f3f77eb6bd43e4f9"},{url:"/_next/static/chunks/app/api/(admins)/reports/tasco/route-74dcba5c104fac3b.js",revision:"74dcba5c104fac3b"},{url:"/_next/static/chunks/app/api/(admins)/reports/truckingReport/route-cb86636a10c194b7.js",revision:"cb86636a10c194b7"},{url:"/_next/static/chunks/app/api/(clock)/getEquipmentList/route-519804720da908d7.js",revision:"519804720da908d7"},{url:"/_next/static/chunks/app/api/(clock)/getLogs/route-a3c44689dc229dcd.js",revision:"a3c44689dc229dcd"},{url:"/_next/static/chunks/app/api/(clock)/getMyTeamsUsers/route-6f02a95895bd7580.js",revision:"6f02a95895bd7580"},{url:"/_next/static/chunks/app/api/(clock)/getPendingTeamTimesheets/%5BcrewMembers%5D/route-49a6fb587a68d661.js",revision:"49a6fb587a68d661"},{url:"/_next/static/chunks/app/api/(clock)/getRecentCostCodes/route-7b2a5a6bdd75cb56.js",revision:"7b2a5a6bdd75cb56"},{url:"/_next/static/chunks/app/api/(clock)/getTodaysTimesheets/route-0e3e4464df7bb4bd.js",revision:"0e3e4464df7bb4bd"},{url:"/_next/static/chunks/app/api/(clock)/getTruckData/route-11b15a94c89f3a17.js",revision:"11b15a94c89f3a17"},{url:"/_next/static/chunks/app/api/(equipment)/getAllEquipmentIdAndQrId/route-a86c0b52bf966423.js",revision:"a86c0b52bf966423"},{url:"/_next/static/chunks/app/api/(equipment)/getCheckedList/route-29cf46270b7c0b16.js",revision:"29cf46270b7c0b16"},{url:"/_next/static/chunks/app/api/(equipment)/getEqUserLogs/%5Bid%5D/route-5031e3c8665a3a8a.js",revision:"5031e3c8665a3a8a"},{url:"/_next/static/chunks/app/api/(equipment)/getEqUserLogs/route-9e1429a0f3ecf97a.js",revision:"9e1429a0f3ecf97a"},{url:"/_next/static/chunks/app/api/(equipment)/getEquipmentRefueledLogs/%5Bid%5D/route-154e76b27348348e.js",revision:"154e76b27348348e"},{url:"/_next/static/chunks/app/api/(equipment)/getRecentJobDetails/route-593c8a1b40146ccc.js",revision:"593c8a1b40146ccc"},{url:"/_next/static/chunks/app/api/(forms)/employeeRequests/%5Bfilter%5D/route-79b1b4688fe82749.js",revision:"79b1b4688fe82749"},{url:"/_next/static/chunks/app/api/(forms)/form/%5Bid%5D/route-5cc9fdcfe1038764.js",revision:"5cc9fdcfe1038764"},{url:"/_next/static/chunks/app/api/(forms)/formDraft/%5Bid%5D/route-38d0f679bc982e5b.js",revision:"38d0f679bc982e5b"},{url:"/_next/static/chunks/app/api/(forms)/formDrafts/route-09288ae8c73e0c79.js",revision:"09288ae8c73e0c79"},{url:"/_next/static/chunks/app/api/(forms)/formSubmission/%5Bid%5D/route-c5c102d4a467d91e.js",revision:"c5c102d4a467d91e"},{url:"/_next/static/chunks/app/api/(forms)/formSubmissions/%5Bstatus%5D/route-7a0273cf39dae2ba.js",revision:"7a0273cf39dae2ba"},{url:"/_next/static/chunks/app/api/(forms)/forms/route-28e2a731008e9bc7.js",revision:"28e2a731008e9bc7"},{url:"/_next/static/chunks/app/api/(forms)/getDocumentById/%5Bid%5D/route-4205af3ca921846b.js",revision:"4205af3ca921846b"},{url:"/_next/static/chunks/app/api/(forms)/getDocumentTags/route-2ab76c44117504b8.js",revision:"2ab76c44117504b8"},{url:"/_next/static/chunks/app/api/(forms)/getDocuments/route-871245f433d876c3.js",revision:"871245f433d876c3"},{url:"/_next/static/chunks/app/api/(forms)/getEmployees/route-351b5ca277a5cc1a.js",revision:"351b5ca277a5cc1a"},{url:"/_next/static/chunks/app/api/(forms)/getEqForDocs/route-739a19704eb7f5a5.js",revision:"739a19704eb7f5a5"},{url:"/_next/static/chunks/app/api/(forms)/managerFormApproval/%5Bid%5D/route-5201e65a13478ba4.js",revision:"5201e65a13478ba4"},{url:"/_next/static/chunks/app/api/(forms)/teamSubmission/%5Bid%5D/route-bb8ab9a66aa355af.js",revision:"bb8ab9a66aa355af"},{url:"/_next/static/chunks/app/api/(general)/cookies/route-f7ea059705edfe5b.js",revision:"f7ea059705edfe5b"},{url:"/_next/static/chunks/app/api/(general)/getAllJobsites/route-5aedb815f8f934ab.js",revision:"5aedb815f8f934ab"},{url:"/_next/static/chunks/app/api/(general)/getAllTags/route-4899eaf2afc5cbb9.js",revision:"4899eaf2afc5cbb9"},{url:"/_next/static/chunks/app/api/(general)/getBannerData/route-253b4ba6c6587580.js",revision:"253b4ba6c6587580"},{url:"/_next/static/chunks/app/api/(general)/getCostCodes/route-32e9164c947e6c1b.js",revision:"32e9164c947e6c1b"},{url:"/_next/static/chunks/app/api/(general)/getCrewByCrewId/%5Bcrew%5D/route-dcf87156946b685e.js",revision:"dcf87156946b685e"},{url:"/_next/static/chunks/app/api/(general)/getEquipment/route-a6dda7b148dac447.js",revision:"a6dda7b148dac447"},{url:"/_next/static/chunks/app/api/(general)/getPayPeriodTimeSheets/route-ecf0ced3fc49b439.js",revision:"ecf0ced3fc49b439"},{url:"/_next/static/chunks/app/api/(general)/getRecentEquipment/route-820910284410c9dc.js",revision:"820910284410c9dc"},{url:"/_next/static/chunks/app/api/(general)/getRecentJobsites/route-8102539df32e6741.js",revision:"8102539df32e6741"},{url:"/_next/static/chunks/app/api/(general)/getRecentTimecard/route-7b01d2b3275d2e61.js",revision:"7b01d2b3275d2e61"},{url:"/_next/static/chunks/app/api/(general)/getRecentTimecardReturn/route-3bbf0b2bd9db103e.js",revision:"3bbf0b2bd9db103e"},{url:"/_next/static/chunks/app/api/(general)/getSettings/route-643152c22a75a21c.js",revision:"643152c22a75a21c"},{url:"/_next/static/chunks/app/api/(general)/getTimesheets/%5Bemployee%5D/route-977fee42de5e0bb1.js",revision:"977fee42de5e0bb1"},{url:"/_next/static/chunks/app/api/(general)/getTimesheets/route-a9c286c949ece94b.js",revision:"a9c286c949ece94b"},{url:"/_next/static/chunks/app/api/(general)/getUserSignature/route-bfc644e772bf5c5f.js",revision:"bfc644e772bf5c5f"},{url:"/_next/static/chunks/app/api/(mechanic)/getMaintenanceLogs/%5Bid%5D/route-3501e832c760ef0e.js",revision:"3501e832c760ef0e"},{url:"/_next/static/chunks/app/api/(mechanic)/getMaintenanceProjects/route-56fd3b1b9f7f257d.js",revision:"56fd3b1b9f7f257d"},{url:"/_next/static/chunks/app/api/(mechanic)/getReceivedInfo/%5Bid%5D/route-a18d16fae8d59d27.js",revision:"a18d16fae8d59d27"},{url:"/_next/static/chunks/app/api/(mechanic)/getRepairDetails/%5Bid%5D/route-ff9286d6db91936a.js",revision:"ff9286d6db91936a"},{url:"/_next/static/chunks/app/api/(myTeam)/getCrew/route-6189b264a6a17479.js",revision:"6189b264a6a17479"},{url:"/_next/static/chunks/app/api/(myTeam)/getCrewById/%5BcrewId%5D/route-c3d9fa8b97fa5c7b.js",revision:"c3d9fa8b97fa5c7b"},{url:"/_next/static/chunks/app/api/(myTeam)/getPendingTeamTimeSheets/route-adaf469f7759eefa.js",revision:"adaf469f7759eefa"},{url:"/_next/static/chunks/app/api/(myTeam)/getTeam/route-3cf21d9cc369f46c.js",revision:"3cf21d9cc369f46c"},{url:"/_next/static/chunks/app/api/(myTeam)/getUserInfo/%5Bid%5D/route-a8b3b9b339d77de8.js",revision:"a8b3b9b339d77de8"},{url:"/_next/static/chunks/app/api/(qrGenerator)/getEquipmentbyQrId/%5BqrId%5D/route-f3750198a0eff189.js",revision:"f3750198a0eff189"},{url:"/_next/static/chunks/app/api/(qrGenerator)/getJobsites/route-1260a25b244737a9.js",revision:"1260a25b244737a9"},{url:"/_next/static/chunks/app/api/(reviewWithDevun)/getSignature/%5Bid%5D/route-0d74e81c1ffdb230.js",revision:"0d74e81c1ffdb230"},{url:"/_next/static/chunks/app/api/(reviewWithDevun)/getTimesheetById/%5Bemployee%5D/route-f0394b3155ea47a1.js",revision:"f0394b3155ea47a1"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/comment/%5BtimeSheetId%5D/route-78e61373d1cab7ce.js",revision:"78e61373d1cab7ce"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/loadCount/%5BtimeSheetId%5D/route-a862ff6accf96c4b.js",revision:"a862ff6accf96c4b"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/refueledLogs/%5BtimeSheetId%5D/route-6896cc408bfead82.js",revision:"6896cc408bfead82"},{url:"/_next/static/chunks/app/api/(tasco)/getTascoLog/tascoId/route-6b984e161dc8640c.js",revision:"6b984e161dc8640c"},{url:"/_next/static/chunks/app/api/(timeSheets)/getComment/route-b96bc0053c8285ca.js",revision:"b96bc0053c8285ca"},{url:"/_next/static/chunks/app/api/(timeSheets)/getMaterialTypes/route-a6e81d8e0f25ce1f.js",revision:"a6e81d8e0f25ce1f"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/endingMileage/%5BtimeSheetId%5D/route-6400ee7d114ab76e.js",revision:"6400ee7d114ab76e"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/equipmentHauled/%5BtimeSheetId%5D/route-dae715faacbdc170.js",revision:"dae715faacbdc170"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/material/%5BtimeSheetId%5D/route-e14f706ef38876ad.js",revision:"e14f706ef38876ad"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/notes/%5BtimeSheetId%5D/route-a8df566facdb87a5.js",revision:"a8df566facdb87a5"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/refueledLogs/%5BtimeSheetId%5D/route-d707528893612245.js",revision:"d707528893612245"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/startingMileage/%5BtimeSheetId%5D/route-73979c38bc3c694a.js",revision:"73979c38bc3c694a"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/stateMileage/%5BtimeSheetId%5D/route-58292ddbc71f8c48.js",revision:"58292ddbc71f8c48"},{url:"/_next/static/chunks/app/api/(trucking)/getTruckingLogs/truckingId/route-96417f5ccf46d783.js",revision:"96417f5ccf46d783"},{url:"/_next/static/chunks/app/api/(user)/getUserImage/route-ead4f465a0cffdf7.js",revision:"ead4f465a0cffdf7"},{url:"/_next/static/chunks/app/api/auth/%5B...nextauth%5D/route-7ccbfd55f62725ba.js",revision:"7ccbfd55f62725ba"},{url:"/_next/static/chunks/app/api/clockoutDetails/route-adc8b0b486305055.js",revision:"adc8b0b486305055"},{url:"/_next/static/chunks/app/api/equipment/%5BequipmentId%5D/lastMileage/route-886eedfd2aa2cce3.js",revision:"886eedfd2aa2cce3"},{url:"/_next/static/chunks/app/api/getTimesheetDetailsManager/%5Bid%5D/route-5465e55fc785b46e.js",revision:"5465e55fc785b46e"},{url:"/_next/static/chunks/app/api/getTimesheetsByDate/route-f5d77af050d64c1e.js",revision:"f5d77af050d64c1e"},{url:"/_next/static/chunks/app/api/getTimesheetsByDateNew/route-f7724fd5ba534f70.js",revision:"f7724fd5ba534f70"},{url:"/_next/static/chunks/app/api/getTruckingLogs/equipmentId/%5BtimeSheetId%5D/route-92e9076b3f9b6302.js",revision:"92e9076b3f9b6302"},{url:"/_next/static/chunks/app/api/sentry-example-api/route-28b9c2ee47fa956d.js",revision:"28b9c2ee47fa956d"},{url:"/_next/static/chunks/app/api/truckingLogDetails/%5BtruckingLogId%5D/route-63b60b14edffe07b.js",revision:"63b60b14edffe07b"},{url:"/_next/static/chunks/app/api/user-permissions/route-3b37bd7da01b67a2.js",revision:"3b37bd7da01b67a2"},{url:"/_next/static/chunks/app/error-8ed5a2d151f85834.js",revision:"8ed5a2d151f85834"},{url:"/_next/static/chunks/app/global-error-293467dba0cb08e6.js",revision:"293467dba0cb08e6"},{url:"/_next/static/chunks/app/icons/page-c8cf36ee1169cb53.js",revision:"c8cf36ee1169cb53"},{url:"/_next/static/chunks/app/layout-9c800248433d75d3.js",revision:"9c800248433d75d3"},{url:"/_next/static/chunks/app/not-found-11295bbc0cdb94b7.js",revision:"11295bbc0cdb94b7"},{url:"/_next/static/chunks/app/page-545d0c878cb6aaa6.js",revision:"545d0c878cb6aaa6"},{url:"/_next/static/chunks/app/sentry-example-page/page-fc2572366eab84f9.js",revision:"fc2572366eab84f9"},{url:"/_next/static/chunks/framework-5871c7b6f4022a8e.js",revision:"5871c7b6f4022a8e"},{url:"/_next/static/chunks/main-38602b92a1f75bdb.js",revision:"38602b92a1f75bdb"},{url:"/_next/static/chunks/main-app-48ed99dcfe0feb7a.js",revision:"48ed99dcfe0feb7a"},{url:"/_next/static/chunks/pages/_app-5376b67503103dac.js",revision:"5376b67503103dac"},{url:"/_next/static/chunks/pages/_error-e94a0482ccaf0c05.js",revision:"e94a0482ccaf0c05"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-4911d3486e27635d.js",revision:"4911d3486e27635d"},{url:"/_next/static/css/63c7182515913e2f.css",revision:"63c7182515913e2f"},{url:"/_next/static/css/ad7c069b3871adaf.css",revision:"ad7c069b3871adaf"},{url:"/_next/static/css/ae3f615d9a81d81a.css",revision:"ae3f615d9a81d81a"},{url:"/_next/static/css/b5fe2f5a95132839.css",revision:"b5fe2f5a95132839"},{url:"/_next/static/css/ccce47fdcedd91d2.css",revision:"ccce47fdcedd91d2"},{url:"/_next/static/media/30e6a981a0efeb9b-s.woff2",revision:"5cd14db2bb47ab5db74376b77c909759"},{url:"/_next/static/media/5df40a3eb8df1a4c-s.p.woff2",revision:"2f46a0d75dfff5f623a08402583e7ee9"},{url:"/_next/static/media/62d466ce3199db61-s.p.woff2",revision:"3a28d888908be7f560bf59faa1c796b2"},{url:"/_next/static/media/92c2336771404627-s.woff2",revision:"e456d9e276ba68ad76f65962e88fab81"},{url:"/_next/static/media/ajax-loader.0b80f665.gif",revision:"0b80f665"},{url:"/_next/static/media/d934919622a522a7-s.woff2",revision:"d3322001406183e84b58b99afd278b84"},{url:"/_next/static/media/slick.25572f22.eot",revision:"25572f22"},{url:"/_next/static/media/slick.653a4cbb.woff",revision:"653a4cbb"},{url:"/_next/static/media/slick.6aa1ee46.ttf",revision:"6aa1ee46"},{url:"/_next/static/media/slick.f895cfdf.svg",revision:"f895cfdf"},{url:"/_next/static/xqew6GGMOd7-5LQPGd-Go/_buildManifest.js",revision:"0e206a604e8ff252314efab19db4eacf"},{url:"/_next/static/xqew6GGMOd7-5LQPGd-Go/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/admin-white.svg",revision:"f60bbe82328c01c11b8e0b2970c889e6"},{url:"/admin.svg",revision:"0d220493364872b69d6f6cde6e385d2c"},{url:"/android/android-launchericon-144-144.png",revision:"554cb47f9b447d7c049e0af337f30598"},{url:"/android/android-launchericon-192-192.png",revision:"c7acce3a1ca251dc273817cd5462a666"},{url:"/android/android-launchericon-48-48.png",revision:"b3e487bbda706b0a7a31bd383fcb5bb9"},{url:"/android/android-launchericon-512-512.png",revision:"daaca9c8424139dcfdeed3181b0a7cbd"},{url:"/android/android-launchericon-72-72.png",revision:"0544816fbb562d0825feb6e9ff73edff"},{url:"/android/android-launchericon-96-96.png",revision:"29c67668a313466cbd66948974b9c656"},{url:"/arrowBack.svg",revision:"17d544aac2962f3f0b83d7dd8bf8d5e7"},{url:"/arrowDown.svg",revision:"bde6e8ea1fa48902e773fc0f34d8ad2c"},{url:"/arrowDownThin.svg",revision:"c8dc31e8dcdca0b4feb827ad7b28c336"},{url:"/arrowLeft.svg",revision:"548bd36e974c05627ee6042a0d17e6c6"},{url:"/arrowLeftSymbol.svg",revision:"4cfdc72b6b5f2cf11ea191dc60c5c4b4"},{url:"/arrowLeftThin.svg",revision:"c4bbdff0ab65d617688561fee993f1c1"},{url:"/arrowRight.svg",revision:"36ae3b90062b7432459cad98be741a80"},{url:"/arrowRightSymbol.svg",revision:"e771082a8a75d50f59df162c67d26091"},{url:"/arrowRightThin.svg",revision:"73a5a9d7e4bf4951d11c4e268faeee85"},{url:"/arrowUp.svg",revision:"3e3a91581069f2b5dc59c7c465255263"},{url:"/arrowUpThin.svg",revision:"ef72218a70f19849149176593f9a8cee"},{url:"/biometrics.svg",revision:"f42bc2055fc95544b3641a2819a084d7"},{url:"/broken.svg",revision:"c62a555a912a5887033a4661236ae31d"},{url:"/calendar.svg",revision:"d6aa1ba02d5a6253fcec8da68242da2a"},{url:"/camera.svg",revision:"4657cc3ac5e2a543f0829e6100446591"},{url:"/cameraFilled.svg",revision:"bcdbce7deaeb9a1d37058d3befea92fc"},{url:"/checkbox.svg",revision:"c2c40521ced6fc4ba77a3865d9dc345f"},{url:"/clock.svg",revision:"eff35cd79029bdd0eeebc39b052f0fbf"},{url:"/clockBreak.svg",revision:"0f0d1da56fbbed48b97b5dbba7d34e65"},{url:"/clockIn.svg",revision:"02e58b9c6f4a129a7026400944bd15b7"},{url:"/clockOut.svg",revision:"a8b9f471daf782532b4006c2cb20b89a"},{url:"/comment.svg",revision:"7bb5715a71c021149b5c3b811aac572b"},{url:"/comments.svg",revision:"29b1a7eb27c708878a454c2eeaad616d"},{url:"/condense-white.svg",revision:"6e9075aa9d7fcb1c242b78610ea9330a"},{url:"/condense.svg",revision:"bc4d6b17d44724df96d6d4a3016072da"},{url:"/drag.svg",revision:"da36f043748f4e8af0e0f5ee17e289a2"},{url:"/dragDots.svg",revision:"c9d25bda626917ea72ca49977f11a016"},{url:"/dragFormBuilder.svg",revision:"bfdedbf1bc6d64afb3642848a542f55d"},{url:"/endDay.svg",revision:"78010192db1eaa29e95ac4d2d9f3705e"},{url:"/equipment-white.svg",revision:"0760c1e58029025d443832cdc33d7402"},{url:"/equipment.svg",revision:"eeb87295d51e91e2adb6483c18bc2452"},{url:"/eraser.svg",revision:"7ed7c543a0ff5530d3c586b368c51315"},{url:"/export-white.svg",revision:"4793d42ec50e11a8586bbc2a5a2f59c8"},{url:"/export.svg",revision:"09f9524501c9e871487735c2b09fcc43"},{url:"/eye-blue.svg",revision:"a851e7f5c035279390dfa1eaa3387e0d"},{url:"/eye.svg",revision:"b7243fcb645ec88a90d9b66fdc2abca3"},{url:"/eyeSlash.svg",revision:"3bb6de382b51b3a333e146bcf8a5d4f5"},{url:"/favicon_io/android-chrome-192x192.png",revision:"07281900d7ffa5690cc984da6e613805"},{url:"/favicon_io/android-chrome-512x512.png",revision:"ee520bf09fd4e101b8952887694905fc"},{url:"/favicon_io/apple-touch-icon.png",revision:"6f682a595255e6aab2bf0f2daa78403a"},{url:"/favicon_io/favicon-16x16.png",revision:"2d00862eea3e0a26607d3d0ce24d1747"},{url:"/favicon_io/favicon-32x32.png",revision:"84f7a4563c6fb8cd905f2dbedcffede8"},{url:"/fileClosed.svg",revision:"bf5a8aeebd62c1736f9e0062a3a271a0"},{url:"/fileOpen.svg",revision:"9ce1ceadae4b4abba237029e0a410d8f"},{url:"/filterDials.svg",revision:"13341b5339e076be7b361c2e09e4976c"},{url:"/filterFunnel.svg",revision:"743604a4e2e50abbb135746948880f65"},{url:"/form-white.svg",revision:"5f85274db63c253161739424aa2fab34"},{url:"/form.svg",revision:"5828ab7b08b01307092a48136ac39cb0"},{url:"/formApproval.svg",revision:"a53cba1fffa0cf55e4dbd86d0345b5e8"},{url:"/formDuplicate.svg",revision:"898bd36ec2c706978dae4e684575ec3e"},{url:"/formEdit.svg",revision:"017c59837c434df47dce4b457e71996e"},{url:"/formInspect-white.svg",revision:"8e48a64eb645a8b5bdc204706e8e6451"},{url:"/formInspect.svg",revision:"f2fcd14456afc03a583c568f4c4da3c8"},{url:"/formList-white.svg",revision:"e1ced4954a119850e3f06e0980062497"},{url:"/formList.svg",revision:"aef95c6f9019999cb9992dd2a21b3975"},{url:"/formSave.svg",revision:"e504c19b5bbbe70b0d2c83c9a73927b8"},{url:"/formSent.svg",revision:"eca064a2079c76fd61ca20528716ceee"},{url:"/formUndo.svg",revision:"91f221fce7b16f414253f9f1873902b2"},{url:"/hauling.svg",revision:"57a718f3ba6755d80106a01a21063d79"},{url:"/haulingFilled.svg",revision:"ebe07b7f9ca47cc45f7fd6b48e6d02ca"},{url:"/header.svg",revision:"479adabc8b70a52625854fba7e433548"},{url:"/home-white.svg",revision:"50f6d70fd4e0a491044340d69138b394"},{url:"/home.svg",revision:"c78e854f7b56bb952a0a22fc1980bc7d"},{url:"/icon-192X192.png",revision:"47aa596be976071dfff9c1e5455ecc80"},{url:"/icon-384x384.png",revision:"99d1987d20b874836d950b95586968ae"},{url:"/icon-512x512.png",revision:"c16feee422d814c946012979c0d02f00"},{url:"/inbox-white.svg",revision:"dfb451099519b395ed3d9b15daad5708"},{url:"/inbox.svg",revision:"03332320e762076df175dde846d62022"},{url:"/inboxFilled.svg",revision:"8b5e6d06de80a9927ae85007277b98a9"},{url:"/information.svg",revision:"3e9a1c81c8fd396fdf6849975d62dbb5"},{url:"/injury.svg",revision:"18121c3e1dc867f5d0a332a7133057f4"},{url:"/ios/100.png",revision:"5222167951b54f250720976e5798bd81"},{url:"/ios/1024.png",revision:"4046bc171d373187ec11f28c06ceb334"},{url:"/ios/114.png",revision:"2aec0491cbf7eb22d34f3079cf2eff54"},{url:"/ios/120.png",revision:"52ac3d673f239d4c822b20c337c73db2"},{url:"/ios/128.png",revision:"bc9dd541b1a0aaff432df4f5b0fb88b4"},{url:"/ios/144.png",revision:"554cb47f9b447d7c049e0af337f30598"},{url:"/ios/152.png",revision:"3085835ae98a412b5f8bb41b3d7fccae"},{url:"/ios/16.png",revision:"876eb7a741ae46958961bf0ed6c30ba0"},{url:"/ios/167.png",revision:"a7230785a2891b4659792d7d6cab6ffb"},{url:"/ios/180.png",revision:"5cdb37c8967955e3a4fae232138f3c03"},{url:"/ios/192.png",revision:"c7acce3a1ca251dc273817cd5462a666"},{url:"/ios/20.png",revision:"8c0f62f1ddd74accb89e318fe432a70f"},{url:"/ios/256.png",revision:"19665fd60a705b586e5cc5c93eddc1f2"},{url:"/ios/29.png",revision:"536c6f0fcf359bba5e0b45c9bb4fc260"},{url:"/ios/32.png",revision:"55a60f12c21ea3fffa61d8f7a34cfc33"},{url:"/ios/40.png",revision:"a38e8ea2895f2cfe54be4ce7465e7952"},{url:"/ios/50.png",revision:"db849275edb671b9ea7c45ef6f4fc489"},{url:"/ios/512.png",revision:"daaca9c8424139dcfdeed3181b0a7cbd"},{url:"/ios/57.png",revision:"8511d6a6a25878417899d2041de166ff"},{url:"/ios/58.png",revision:"001948872d9d885d20140b7aa9cc6ff8"},{url:"/ios/60.png",revision:"80209b85fb0be10562bcee31908d2a95"},{url:"/ios/64.png",revision:"066fa2094719e23427225f5517889ed7"},{url:"/ios/72.png",revision:"0544816fbb562d0825feb6e9ff73edff"},{url:"/ios/76.png",revision:"0a1c5da7e5a9673345c44f7e1c87f0d4"},{url:"/ios/80.png",revision:"91b74cbd7d0f0119a164a0750462149e"},{url:"/ios/87.png",revision:"74c6a6624174372f1fc483515c820c2f"},{url:"/jobsite-white.svg",revision:"1d991bcf439a2538ed78c709c708954b"},{url:"/jobsite.svg",revision:"f3909b4ff16c4f25e211fb0dc7556050"},{url:"/key.svg",revision:"f39b7da7fb64da838435e228ca3113ed"},{url:"/language.svg",revision:"e4615572d416d5df80c463dc53d5e0b7"},{url:"/layout.svg",revision:"a3b50fb72e8fab8feac60ccbd864e31a"},{url:"/logo.svg",revision:"8e78aff787124cc6567c2ccfd520099d"},{url:"/logo_JPG.jpg",revision:"e84c38cbd142e140688beacd59e6a0cf"},{url:"/manifest.json",revision:"05d4e6a0d9c7c7bebec473d08ea582b2"},{url:"/mechanic.svg",revision:"4f39d9d1ff5c68a81b93c7f6611d8e9d"},{url:"/message.svg",revision:"c55f300ce914f4b28036e30580ee104c"},{url:"/mileage.svg",revision:"11e1eb1d01c9be54e3c8394e198db717"},{url:"/minus.svg",revision:"d2ec83afd3490cac9da747386df54184"},{url:"/moreOptions.svg",revision:"d3019cfec5d1102ec73147f1dd5e5faf"},{url:"/moreOptionsCircle.svg",revision:"02d86b2a2acd7707b4cef84da6ede884"},{url:"/number.svg",revision:"6f90bfd3f431fe8602537fe47b9a109b"},{url:"/plus-stroke-white.svg",revision:"f23aeed115ad4273e185d3e73a314ba7"},{url:"/plus-white.svg",revision:"6adf4b80867ce618d5ca70765c2e429d"},{url:"/plus.svg",revision:"15fadbc574e693de5d7b486f4e13b1d5"},{url:"/policies.svg",revision:"25de43083a5e24cdf83a7ba62d6c28c4"},{url:"/priorityDelay.svg",revision:"5286ef619f78a44968b600b83bba55a0"},{url:"/priorityHigh.svg",revision:"fa44198aecf39d4748914d7f3b4a0d03"},{url:"/priorityLow.svg",revision:"11efcad98767f7c049d90ece2f425f1a"},{url:"/priorityMedium.svg",revision:"219c8a19b23f94db24162d544e1605ec"},{url:"/priorityPending.svg",revision:"45d3ef273daf2b4daadcb0b6d0064789"},{url:"/priorityToday.svg",revision:"91fec1dfbc8a1268488c12a9ef9fa4ac"},{url:"/profileEmpty.svg",revision:"66491e65dd6114a9acf6990181fc44d9"},{url:"/profileFilled.svg",revision:"ff2e851ef303666374501b541725d34c"},{url:"/qrCode-white.svg",revision:"0bcc19567973489fa8a15e759b729d5b"},{url:"/qrCode.svg",revision:"a97e34eddfd5bae153771e734fd96e4c"},{url:"/radio.svg",revision:"17ecc9b4e325da65fc947c71d4bc1ade"},{url:"/refuel.svg",revision:"07fcabd768272b84f61007c4e130b8fc"},{url:"/refuelFilled.svg",revision:"d75e2ee23202812499b36b050fff841c"},{url:"/searchLeft.svg",revision:"7a4c52a1de03e0fbe1f2f949fed97367"},{url:"/searchRight.svg",revision:"9ac605dce8646197fa3ae4c90d93f8c9"},{url:"/settingsFilled.svg",revision:"4b9876c68beb7435b569e5a5a0252501"},{url:"/shiftScanLogo.svg",revision:"f6cdbf4c760167995264751f68ebb2af"},{url:"/shiftscanlogoHorizontal.svg",revision:"f6dac5483d88c1cdebc5f92190cc1d2f"},{url:"/spinner.svg",revision:"7eedb8067da00e29baf517b3cbfffc43"},{url:"/star.svg",revision:"744181e34e34c3385a845fa37cf51454"},{url:"/starFilled.svg",revision:"5ad7cd14d1deca592924c230b5a39218"},{url:"/state.svg",revision:"83c029784c461361996615f83c3850ef"},{url:"/stateFilled.svg",revision:"183c533850d5a2540b71b2c44bef9d11"},{url:"/statusApproved.svg",revision:"0e767f6c9f2c8f20480d3120d6b6e198"},{url:"/statusApprovedFilled.svg",revision:"6fc76b4c67d20fa9ffbdb451a30e8071"},{url:"/statusDenied.svg",revision:"7b1167c6857f8169396ae88ffbce95e6"},{url:"/statusDeniedFilled.svg",revision:"52740140811df8fc4cd87604621c9960"},{url:"/statusOffline.svg",revision:"d1931f81ba9fe9342e78e181c2216aaa"},{url:"/statusOngoing-white.svg",revision:"d07cb3a9608fbfa3e64486e49eb74365"},{url:"/statusOngoing.svg",revision:"7eedb8067da00e29baf517b3cbfffc43"},{url:"/statusOngoingFilled.svg",revision:"5e18aea67f178cddbdc7e2e449ea6e3a"},{url:"/statusOnline.svg",revision:"ebed6ae0692535bcfee10ada59c779c3"},{url:"/statusUnfinished.svg",revision:"038f7353e12e76758c9e214925a07a0c"},{url:"/swe-worker-5c72df51bb1f6ee0.js",revision:"447aee4a2c24fefe7917563f0255d688"},{url:"/tasco.svg",revision:"1280d07dd40b7604d022f71c3728ccfb"},{url:"/team-white.svg",revision:"0142ac2317a2b44b141c270916ab32ca"},{url:"/team.svg",revision:"203a48a7519f73fe7f24fff1cae20204"},{url:"/timecards.svg",revision:"7771f6f39b0e3d775e88c6d542d2d68b"},{url:"/tinyCheckMark-white.svg",revision:"a6d6eadc3dbfaafbf1e8257346eee02c"},{url:"/title.svg",revision:"15b1291ddf60f0ceb36ce46377fa6dd3"},{url:"/trash-red.svg",revision:"a42dfceb5b388b58d0c41c13ee0c17f7"},{url:"/trash.svg",revision:"229905df8ec07dd400f7a23deb582f0c"},{url:"/trucking.svg",revision:"ab3794a1a3073bc2926a1978a038300c"},{url:"/windows11/LargeTile.scale-100.png",revision:"e3ae3637973ea93ff81e73ee37effe0c"},{url:"/windows11/LargeTile.scale-125.png",revision:"6e4ec698528b71f25efd6ff2dd68e89f"},{url:"/windows11/LargeTile.scale-150.png",revision:"b636fa1812d337a4ca1224b433a3d855"},{url:"/windows11/LargeTile.scale-200.png",revision:"d18a045105d03ba23b624c109708f7bb"},{url:"/windows11/LargeTile.scale-400.png",revision:"d4176f9ebaacc0bbfebf56fd56e253ce"},{url:"/windows11/SmallTile.scale-100.png",revision:"bce1445e4fb91c8c0842780a4026522e"},{url:"/windows11/SmallTile.scale-125.png",revision:"9415b7ff287961c074dee1834b42aad9"},{url:"/windows11/SmallTile.scale-150.png",revision:"87d19729c39b8d355a62d180d91bb31b"},{url:"/windows11/SmallTile.scale-200.png",revision:"7c5db43e6077ca3c958c21c85bec6d25"},{url:"/windows11/SmallTile.scale-400.png",revision:"1913d6ed56bc5b31dc60858acbcace6f"},{url:"/windows11/SplashScreen.scale-100.png",revision:"a8215f24409e6231d1dc00a0907ed4a8"},{url:"/windows11/SplashScreen.scale-125.png",revision:"097ac942f0459d3481fbd62c0f97e299"},{url:"/windows11/SplashScreen.scale-150.png",revision:"201fc53647b740af27d9b004cafeaaf8"},{url:"/windows11/SplashScreen.scale-200.png",revision:"1b1081f94203f3823ed2dd1f8dabb29a"},{url:"/windows11/SplashScreen.scale-400.png",revision:"0d6ee6f415a7d2101792e2b8312af059"},{url:"/windows11/Square150x150Logo.scale-100.png",revision:"7b6855c0928dad5db20403341b059afc"},{url:"/windows11/Square150x150Logo.scale-125.png",revision:"dd8460f97de8d2cf5826468f8f8153db"},{url:"/windows11/Square150x150Logo.scale-150.png",revision:"8bfd4727bbeb856bd1af17e0cdef44ea"},{url:"/windows11/Square150x150Logo.scale-200.png",revision:"c95f1e579117442bcf6a388818260303"},{url:"/windows11/Square150x150Logo.scale-400.png",revision:"a20add32148ad6bfcc0d435ac4986eb5"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.altform-lightunplated_targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.altform-unplated_targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/Square44x44Logo.scale-100.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.scale-125.png",revision:"95498060191b331348eb8f5b5ab6dfe7"},{url:"/windows11/Square44x44Logo.scale-150.png",revision:"aced07a7a4627cdad487ed0a46faeebb"},{url:"/windows11/Square44x44Logo.scale-200.png",revision:"db11fa1c0b22e2344ecf9dea9af68a2c"},{url:"/windows11/Square44x44Logo.scale-400.png",revision:"5f1b23fb222c880bf3c533643f43b4ca"},{url:"/windows11/Square44x44Logo.targetsize-16.png",revision:"90a9d6c44b09d44f416b62df26224428"},{url:"/windows11/Square44x44Logo.targetsize-20.png",revision:"779afd7d8c738f0a0c73a4a6d63e4aff"},{url:"/windows11/Square44x44Logo.targetsize-24.png",revision:"a65f2b674acb5c2af36db18e9b4a63f1"},{url:"/windows11/Square44x44Logo.targetsize-256.png",revision:"723b833ab63af25543dfe92f0a56e6d5"},{url:"/windows11/Square44x44Logo.targetsize-30.png",revision:"3a63be52b054d426a0ee4ce69b783e12"},{url:"/windows11/Square44x44Logo.targetsize-32.png",revision:"261d9c7ca92fefc2230721bb99cbcae1"},{url:"/windows11/Square44x44Logo.targetsize-36.png",revision:"a709411193e6374836f5fa82b8afa6c1"},{url:"/windows11/Square44x44Logo.targetsize-40.png",revision:"4251a65706de0f7c86b1bee205eb53a7"},{url:"/windows11/Square44x44Logo.targetsize-44.png",revision:"9f4cf9e0ba353f1bbb19fc9f6e40523a"},{url:"/windows11/Square44x44Logo.targetsize-48.png",revision:"c87a2b33f89afc0199219ea0b6e191ac"},{url:"/windows11/Square44x44Logo.targetsize-60.png",revision:"727adb65708020078e70655ef5444a7b"},{url:"/windows11/Square44x44Logo.targetsize-64.png",revision:"04915cfcb9f9ad29a6223fb193c79a14"},{url:"/windows11/Square44x44Logo.targetsize-72.png",revision:"1f5e7f8fb46e7d8e425ff06c34338b6d"},{url:"/windows11/Square44x44Logo.targetsize-80.png",revision:"3a7aa9e5f873375880d85aa0c2f51306"},{url:"/windows11/Square44x44Logo.targetsize-96.png",revision:"a151f9aefbcedac0d11aba5cf1108af8"},{url:"/windows11/StoreLogo.scale-100.png",revision:"db849275edb671b9ea7c45ef6f4fc489"},{url:"/windows11/StoreLogo.scale-125.png",revision:"728e6527e2d77371e3007caf007b116f"},{url:"/windows11/StoreLogo.scale-150.png",revision:"9224f9a45257c2103e10aa9026af46d1"},{url:"/windows11/StoreLogo.scale-200.png",revision:"5222167951b54f250720976e5798bd81"},{url:"/windows11/StoreLogo.scale-400.png",revision:"6562c92cea67a6232c5727e464a9dd83"},{url:"/windows11/Wide310x150Logo.scale-100.png",revision:"7fc14247c4af58216f16d5f4f98fc46b"},{url:"/windows11/Wide310x150Logo.scale-125.png",revision:"a0c8d4ccacb2bc65438458df95ab3d60"},{url:"/windows11/Wide310x150Logo.scale-150.png",revision:"34161357c0cdafe2651b8d197c019798"},{url:"/windows11/Wide310x150Logo.scale-200.png",revision:"a8215f24409e6231d1dc00a0907ed4a8"},{url:"/windows11/Wide310x150Logo.scale-400.png",revision:"1b1081f94203f3823ed2dd1f8dabb29a"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https?.*/,new e.NetworkFirst({cacheName:"offlineCache",plugins:[new e.ExpirationPlugin({maxEntries:200,maxAgeSeconds:86400})]}),"GET"),self.__WB_DISABLE_DEV_LOGS=!0}));
 //# sourceMappingURL=sw.js.map
